==> src/1_preprocessing.Rmd <==
# Import and preprocess INBO color ring data

## Extract data from INBO kleurring database

### Connection to INBO database

Get connection settings from `config.yml` (not included to the reposiory) and connect to database:

```{r connect_to_db}
meeuwen <- config::get("meeuwen")
conn <- dbConnect(odbc::odbc(),
  driver = meeuwen$driver,
  server = meeuwen$server,
  database = meeuwen$database,
  port = meeuwen$port,
  trusted_connection = meeuwen$trusted_connection
)
```

### Extract user data

First, we extract data about users from INBO *kleurring* database:

```{r get_user_data}
users <- dbGetQuery(conn, "SELECT * FROM dbo.tblWaarnemer") %>%
  as_tibble()
```

### Extract color ring data

```{r get_birds_data}
birds <- dbGetQuery(conn, "SELECT * FROM dbo.tblKleurring") %>%
  as_tibble()
```

We extract the color table in order to know the meaning of the color abbreviations in fields  `RingKleurCode` and `InscriptieKleurCode` in `birds`.

```{r get_color_table}
color_table <- dbGetQuery(conn, "SELECT * FROM dbo.cdeKleur") %>%
  as_tibble()
color_table
```

We also extract the ring position (left or right leg) and the inscription orientation:

```{r  get_ring_position_table}
ring_position_table <- dbGetQuery(conn, "SELECT * FROM dbo.cdeRingPlaats") %>%
  as_tibble()
ring_position_table
```

### Extract observation data

INBO's observation data contain a text type field: `Opmerking`. Text type is deprecated and an error will be returned if we perform the standard SQL query `"SELECT * FROM dbo.tblWaarneming"`. So, we need an ad-hoc query:

```{r get_obs_data}
obs <- dbGetQuery(
  conn,
  "SELECT Nummer,
          Datum,
          EuringCode,
          LeeftijdCode,
          KleurringNummer,
          KleurringPlaats,
          MetaalringNummer,
          MetaalringPlaats,
          PlaatsGemeente,
          PlaatsToponym,
          PlaatsToponymDetail,
          Convert(nvarchar(4000),Opmerking) as Opmerking,
          WaarnemerNummer,
          PlaatsLengtegraadGraden,
          PlaatsLengtegraadMinuten,
          PlaatsLengtegraadSeconden,
          PlaatsBreedtegraadGraden,
          PlaatsBreedtegraadMinuten,
          PlaatsBreedtegraadSeconden,
          PlaatsLengtegraadRichtingCode,
          PlaatsBreedtegraadRichtingCode,
          PlaatsLandCode,
          MetaalringLandCode,
          BevestigingDatum,
          PlaatsProvincie,
          AanmaakDatum,
          WijzigDatum
  FROM dbo.tblWaarneming"
) %>% as_tibble()
```

Table `tblWaarnemingAktie` is also important because contains informations about the actions taken at each observation:

```{r get_waarnemingAktie}
obs_actions <- dbGetQuery(conn, "SELECT * FROM dbo.tblWaarnemingAktie")
```

The action codes are described in table `dbo.cdeAktie`:

```{r get_}
actions_meaning <- dbGetQuery(conn, "SELECT * FROM dbo.cdeAktie")
actions_meaning
```

Close connection to server:

```{r close_connection}
dbDisconnect(conn)
```

### Import control data

From UVABIT repository, we copied a text file containing the mapping of GPS ids:

```{r import_GPS_names_ids}
map_gps_path <- here::here("data", "input", "map_gps_id_to_color_ring.tsv")
map_gps <-
  read_tsv(map_gps_path) %>%
  mutate(sovon_bird_notes = "uvabits_gps_tag.")
map_gps %>% head()
```

This file will be used later in this same document.

# Pre-processing

## Preview raw data

For privacy reasons users data cannot be shown. Users data refer to the following fields:

```{r prview_users}
colnames(users)
```

Preview birds data:

```{r preview_birds}
birds %>% head()
```

Preview observations data:

```{r preview_obs}
obs %>% head()
```

Preview observation actions data:

```{r preview_obs_actions}
obs_actions %>% head()
```

## Check primary key consistency

The primary key of each table shuold never be left empty and should contain unique values.

### Users

Primary key: column `Nummer`. Users with empty values:

```{r check_primary_key_users}
users %>%
  filter(is.na(Nummer))
```

Users with not unique values of `Nummer`:

```{r check_unicity_prkey_users}
users %>%
  group_by(Nummer) %>%
  count() %>%
  filter(n > 1)
```

### Birds

Primary key: column `Nummer`. Birds with empty values:

```{r check_primary_key_birds}
birds %>%
  filter(is.na(Nummer))
```

Birds with not unique values of `Nummer`:

```{r check_unicity_prkey_birds}
birds %>%
  group_by(Nummer) %>%
  count() %>%
  filter(n > 1)
```

### Observations

Primary key: column `Nummer`. Observations with empty values:

```{r check_primary_key_obs}
obs %>%
  filter(is.na(Nummer))
```

Observations with not unique values of `Nummer`:

```{r check_unicity_prkey_obs}
obs %>%
  group_by(Nummer) %>%
  count() %>%
  filter(n > 1)
```

### Actions

Primary key: column `Nummer`. Actions with empty values:

```{r check_primary_key_actions}
obs_actions %>%
  filter(is.na(Nummer))
```

Observations with not unique values of `Nummer`:

```{r check_unicity_prkey_actions}
obs_actions %>%
  group_by(Nummer) %>%
  count() %>%
  filter(n > 1)
```

## Remove hard returns

### Remove hard returns in `users`

Remove hard returns `\r\n` in `users`:

```{r remove_hard_return_users}
users <-
  users %>%
  mutate_if(is.character, ~ str_replace_all(., "[\\r\\n]", ""))
```

### Remove hard returns in `birds`

Remove hard returns `\r\n` in `birds`:

```{r remove_hard_return_tabs_birds}
birds <-
  birds %>%
  mutate_if(is.character, ~ str_replace_all(., "[\\r\\n]", ""))
```

### Remove hard returns in `observations`

Remove hard returns `\r\n` in `observations`:

```{r remove_hard_return_obs}
obs <-
  obs %>%
  mutate_if(is.character, ~ str_replace_all(., "[\\r\\n]", ""))
```

## Check spaces in `birds`

No spaces should be present in any column of `birds`:

```{r check_spaces}
map_dfr(birds, ~ mutate(birds, space_detect = str_detect(., pattern = " "))) %>%
  filter(space_detect == TRUE)
```

Remove all spaces:

```{r remove_spaces}
birds <- birds %>%
  mutate_all(list(~ str_replace_all(., " ", "")))
```

## Check duplicates `WaarnemingNummer`-`AktieCode`

The action acronym is contained in column `AktieCode`. Observations and their actions are linked via columns `WaarnemingNummer` (in `actions`) and `Nummer` (in `obs`).

No duplicates `WaarnemingNummer`-`AktieCode` should exist:

```{r show_duplicates_actions}
obs_actions %>%
  group_by(WaarnemingNummer, AktieCode) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
```

Otherwise they should be removed:

```{r to_be_removed}
obs_actions <-
  obs_actions %>%
  distinct(WaarnemingNummer, AktieCode, .keep_all = TRUE)
```

## Merge observations and actions

In order to ease the mapping of birds and observations, we first remove columns will be never used: 

```{r remove_cols_not_used}
obs_actions <-
  obs_actions %>%
  select(-c(
    Nummer,
    AanmaakDatum,
    WijzigDatum
  ))
```

Then we *spread* action codes to columns:

```{r spread_action_codes}
obs_actions <-
  obs_actions %>%
  pivot_wider(names_from = AktieCode, values_from = AktieCode)
```

The following combinations occur:

```{r distinct_combinations_actions}
combinations_actions <- 
  obs_actions %>%
  select(-WaarnemingNummer) %>%
  distinct()
combinations_actions
```

Actions present in database:

```{r colnames_actions_present}
acts <- names(combinations_actions)
acts
```

Actions described in `actions_meaning` never used:

```{r no_used_actions}
actions_meaning %>%
  filter(!Code %in% acts)
```

We add action code columns to observations in order to have a unique dataframe for observations:

```{r merge_obs_with_acts}
obs_and_acts <-
  obs %>%
  left_join(obs_actions, by = c("Nummer" = "WaarnemingNummer"))
```

Preview: 

```{r merge_obs_with_acts_preview}
head(obs_and_acts, n = 100)
```

## Check spaces in `obs_and_acts`

No spaces should be present in character columns of `obs_and_acts` except for columns related to place description (`Plaats*`), datums (`*Datum`) and notes (`Opmerking`):

```{r show_spaces_obs_and_acts}
map_dfr(obs_and_acts %>%
  select_if(is.character) %>%
  select(-c(
    starts_with("Plaats"),
    Opmerking,
    ends_with("Datum")
  )), ~
mutate(obs_and_acts, space_detect = str_detect(., pattern = " "))) %>%
  filter(space_detect == TRUE) %>%
  arrange(Nummer)
```

Remove spaces if present:

```{r remove_spaces_from_obs_and_acts}
obs_and_acts <-
  obs_and_acts %>%
  mutate(across(
    !starts_with("Plaats") &
      !ends_with("Datum") & 
      !one_of("Opmerking"),
    ~ str_replace_all(., " ", "")
  ))
```

## Check presence action code

All observations should also have an action code:

```{r observations_without_action}
no_actions <-
  obs_and_acts %>%
  filter_at(vars_select(names(obs_and_acts), one_of(acts)), all_vars(is.na(.)))
no_actions
```

Observation without it will be removed:

```{r remove_observations_without_action}
obs_and_acts <-
  obs_and_acts %>%
  filter(!Nummer %in% no_actions$Nummer)
```

## Check combinations of actions with `klgev` or `megev`

`klgev` should never occur with other *active* actions like `vang`, `vangl` or action `dood`. Is this true?

```{r check_acts_with_klgev}
acts_with_klgev <-
  obs_and_acts %>%
  filter(klgev == "klgev") %>%
  select(all_of(acts)) %>%
  distinct() %>%
  select_if(~ sum(!is.na(.)) > 0) %>%
  names()
acts_with_klgev <- acts_with_klgev[acts_with_klgev != "klgev"]
length(acts_with_klgev) == 0
```

Same for `megev`:

```{r check_acts_with_megev}
if ("megev" %in% acts) {
  acts_with_megev <-
    obs_and_acts %>%
    filter(megev == "megev") %>%
    select(all_of(acts)) %>%
    distinct() %>%
    select_if(~ sum(!is.na(.)) > 0) %>%
    names()
  acts_with_megev <- acts_with_megev[acts_with_megev != "megev"]
  length(acts_with_megev) == 0
} else {
  message("No 'megev' actions found.")
}
```

## Remove data labelled as `FOUT`

There are observations judged as error. They are marked by`KleurringNummmer` equal to `FOUT`:

```{r show_FOUT_obs}
obs_and_acts %>%
  filter(KleurringNummer == "FOUT")
```

The value `FOUT` is also present in `birds`:

```{r show_FOUT_birds}
birds %>%
  filter(Nummer == "FOUT")
```

These data are related to observations judged not correct (`FOUT`: _error_) by the INBO experts and administrator of the database.

Thus, at the moment, we remove these data (discussion about this issue [here](https://github.com/inbo/sovon/issues/9)) from observations:

```{r remove_FOUT_obs}
obs_and_acts <-
  obs_and_acts %>%
  filter(KleurringNummer != "FOUT")
```

and from birds: 

```{r remove_FOUT_birds}
birds <-
  birds %>% filter(Nummer != "FOUT")
```

## Remove `EAYT`

The bird with color ring `EAYT` has been added after the final data export for SOVON, Oct 8th, and will be removed from birds:

```{r remove_EAYT}
birds <-
  birds %>%
  filter(Nummer != "EAYT")
```

and observations:

```{r remove_EAYT_from_obs}
obs_and_acts <-
  obs_and_acts %>%
  filter(KleurringNummer != "EAYT")
```

## Check color rings: uppercase

Bird color rings codes are uppercase. Exceptions:

```{r show_rings_lowercase}
birds %>%
  filter(Nummer != toupper(Nummer) |
    NummerNieuw != toupper(NummerNieuw) |
    NummerDesc != toupper(NummerDesc))
```

are converted to uppercase:

```{r convert_rings_to_uppercase}
birds <- birds %>%
  mutate(
    Nummer = toupper(Nummer),
    NummerNieuw = toupper(NummerNieuw),
    NummerDesc = toupper(NummerDesc)
  )
```

The same holds true for `KleurringNummer` in `obs_and_acts`. Exceptions:

```{r lowercase_kleurringNummer}
obs_and_acts %>%
  filter(KleurringNummer != toupper(KleurringNummer)) %>%
  distinct(KleurringNummer)
```

are transformed to uppercase:

```{r KleurringNummer_to_upper}
obs_and_acts <- obs_and_acts %>%
  mutate(KleurringNummer = toupper(KleurringNummer))
```

## Check keys `KleurringNummer` (obs) - `Nummer` (birds)

All values in `KleurringNummer` should be present in `birds$Nummer`. Exceptions:

```{r KleurringNummer_not_in_birds_Nummer}
obs_and_acts %>%
  filter(!KleurringNummer %in% birds$Nummer) %>%
  distinct(KleurringNummer)
```

We remove them:

```{r remove_obs_and_acts_linked_to_invalid_bird_ring}
KleurringNummer_remove <-
  obs_and_acts %>%
  filter(!KleurringNummer %in% birds$Nummer) %>%
  distinct(KleurringNummer) %>%
  pull()
obs_and_acts <-
  obs_and_acts %>%
  filter(!KleurringNummer %in% KleurringNummer_remove)
```

Finally, we search for birds (`Nummer` in  `birds`) not linked to any observation (`KleurringNummer` in `obs_and_acts`):

```{r birds_not_in_obs}
birds %>% filter(!Nummer %in% obs_and_acts$KleurringNummer)
```

We remove them:

```{r remove_birds_not_in_obs}
birds <-
  birds %>%
  filter(Nummer %in% obs_and_acts$KleurringNummer)
```

## Check duplicates in date of applying rings

Detect duplicates in date of applying (the very first) color ring (action code `rngkl`):

```{r duplicates_apply_first_color_ring}
duplicates_ringing_rngkl <-
  obs_and_acts %>%
  filter(!is.na(rngkl)) %>%
  group_by(KleurringNummer, Datum, rngkl) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
duplicates_ringing_rngkl
```

and metal ring (action code `rngme`):

```{r duplicates_apply_first_metal_ring}
duplicates_ringing_rngme <-
  obs_and_acts %>%
  filter(!is.na(rngme)) %>%
  group_by(KleurringNummer, Datum, rngme) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
duplicates_ringing_rngme
```

## Add `rngme` if not present

Some metal rings are not linked to an observation with action `rngme`, so we have no date for them:

```{r check_apply_rngme}
no_rngme <-
  birds %>%
  filter(!is.na(MetaalringNummer)) %>%
  left_join(obs_and_acts %>%
    filter(!is.na(rngme)) %>%
    select(KleurringNummer, Datum),
  by = c("Nummer" = "KleurringNummer")
  ) %>%
  filter(is.na(Datum)) %>%
  select(Nummer, MetaalringNummer) %>%
  rename(KleurringNummer = Nummer) %>%
  arrange(KleurringNummer)
no_rngme
```

Experts are sure that these metal rings have been applied while applying the color ring (see [comment on issue #33](https://github.com/inbo/sovon/issues/33#issuecomment-451876667)). So, we can add action `rngme` to observations coupled to action `rngkl` for these rings:

```{r add_rngme_to_obs_with_rngkl}
obs_and_acts <-
  obs_and_acts %>%
  left_join(no_rngme %>%
    select(KleurringNummer) %>%
    mutate(add_rngme = TRUE),
  by = "KleurringNummer"
  ) %>%
  mutate(rngme = case_when(
    rngkl == "rngkl" & add_rngme == TRUE ~ "rngme",
    TRUE ~ rngme
  )) %>%
  select(-add_rngme)
```

Be sure this solution provides a date for all metal rings:

```{r check_date_for_all_metal_rings}
obs_and_acts %>%
  filter(is.na(rngme) | rngme == "rngme") %>%
  nrow() == nrow(obs_and_acts)
```

## Check age while applying rings

For each bird, observations with same date should refer to same bird age. The only exceptions allowed are those from  `veld` observations as the age in this case is estimated by observer and can be different from the real one.  
Exceptions, afer removing field observations:

```{r check_anomalies_age_ringing}
exceptions_age <-
  obs_and_acts %>%
  # age is present
  filter(!is.na(LeeftijdCode)) %>%
  # exclude field observations
  filter(is.na(veld)) %>%
  distinct(KleurringNummer, Datum, LeeftijdCode, WaarnemerNummer) %>%
  group_by(KleurringNummer, Datum, WaarnemerNummer) %>%
  count() %>%
  filter(n > 1) %>%
  left_join(obs_and_acts %>%
    filter(!is.na(LeeftijdCode)) %>%
    select(
      Nummer, KleurringNummer, Datum, WaarnemerNummer,
      LeeftijdCode, one_of(actions_meaning$Code)
    )) %>%
  select(-n) %>%
  select(
    Nummer, KleurringNummer, Datum,
    LeeftijdCode, WaarnemerNummer, one_of(actions_meaning$Code)
  )
exceptions_age
```

## Check bird sex inconsistencies

All birds should have a sex, which could be one of `M` (_mannetje_), `V` (_vrouwtje_), `O` (_onbekend_):

```{r present_GeslachtCode}
distinct(birds, GeslachtCode)
```

Exceptions:

```{r exceptions_bird_sex}
birds %>%
  filter(is.na(GeslachtCode)) %>%
  distinct(Nummer, NummerNieuw, NummerDesc, GeslachtCode)
```

For all birds a one-to-one relation bird - sex should hold true:

```{r check_one-to-one_bird_sex}
birds %>%
  distinct(Nummer, GeslachtCode) %>%
  nrow() == nrow(birds)
```

## Handle information about metal ring number

There are columns containing metal ring related information in `birds`:

```{r cols_metal_ring_birds}
names(birds)[which(str_detect(names(birds), "Metaalring"))]
```

and `obs_and_acts`:

```{r cols_metal_ring_obs_and_acts}
names(obs_and_acts)[which(str_detect(names(obs_and_acts), "Metaalring"))]
```

This is not only redundant: inconsistencies are detected. Some examples:

```{r inconsisntecy_metal_ring_info_examples}
birds %>%
  filter(Nummer %in% c("AAAA", "AAAR", "AAAT", "AAAW")) %>%
  rename("metal_ring_from_birds" = "MetaalringNummer") %>%
  left_join(obs_and_acts %>%
    rename("metal_ring_from_obs" = "MetaalringNummer"),
  by = c("Nummer" = "KleurringNummer")
  ) %>%
  distinct(Nummer, metal_ring_from_birds, metal_ring_from_obs)
```

Based on experts' knowledge only the metal ring information in `birds` is correct. We rename the columns related to metal ring information from `obs_and_acts` by adding suffix `_obs`:

```{r remove_metal_ring_infos_from_obs}
obs_and_acts <-
  obs_and_acts %>%
  rename_at(
    vars(starts_with("Metaalring")),
    ~ paste0(., "_obs")
  )
```

## Solve GPS tracker IDs

Some birds have a suspect `Nummer` which is formed by letters `GPS` or `GP` followed by some numbers. These are a kind of GPS IDs which should not be in columns related to color rings:

```{r show_gps_ids}
birds %>%
  filter(str_detect(Nummer, pattern = "(^(GP)\\d+)|(^(GPS)\\d+)")) %>%
  select(Nummer)
```

The right mapping is saved in `map_gps`:

```{r preview_map_gps}
map_gps %>% head()
```

We import the color rings in new column `first_Nummer`. For all other rings will be `first_nummer = Nummer` as they are the very first rings:

```{r map_to_first_Nummer}
birds <-
  birds %>%
  # create column first_Nummer with right color rings for the GPS or GP** rings
  left_join(map_gps, by = c("Nummer" = "gps_id")) %>%
  # set first_Numer equal to Nummer in all other cases
  mutate(first_Nummer = ifelse(is.na(color_ring),
    Nummer,
    color_ring
  )) %>%
  select(Nummer, first_Nummer, everything())
```

Check whether not corrected rings (`GP***` or `GPS***`) are still present:

```{r check_presence_GP_GPS_in_birds}
birds %>%
  filter(str_detect(first_Nummer, pattern = "(^(GP)\\d+)|(^(GPS)\\d+)")) %>%
  select(first_Nummer, NummerNieuw)
```

Mapping example: the bird with `Nummer` equal to `GP104` is mapped as follows:

```{r example_gps_mapping_and_name}
birds %>%
  filter(Nummer == "GP104") %>%
  select(Nummer, first_Nummer, NummerNieuw, sovon_bird_notes)
```

Column `first_Nummer` should contain unique elements. Check for exceptions:

```{r check_uniqueness_first_Nummer}
birds %>%
  group_by(first_Nummer) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1) %>%
  select(-n)
```

Exceptions, if exist, are due to double entry, one with the original color ring, a second one with GPS tracker ID. Such exceptions, if present, should be handled separately while mapping `bird_shorthand` in document `3_birds.Rmd`. Another reason could be mapping in `map_gps` as result of reringing of birds with GPS. Notice that `map_gps` should map to the very first ring.

## Check inconsistencies between users and observers

We have to be sure that all observations have a valid observator reference, i.e. an ID contained in column `Nummer` of `users`.

Observations without an observator reference:

```{r WaarnemerNummer_is_na}
obs_and_acts %>%
  filter(is.na(WaarnemerNummer))
```

All observator references in `obs_and_acts` should be present in `users`:

```{r check_observators_in_users}
observers_ids <- unique(
  obs_and_acts %>%
    filter(!is.na(WaarnemerNummer)) %>%
    distinct(WaarnemerNummer) %>%
    pull()
)
all(observers_ids %in% users$Nummer)
```

## Remove data of common shelduck

Data related to common shelduck, *Tadorna tadorna*, are not updated: INOB experts suggest to remove them from the mapping. More details [here](https://github.com/inbo/sovon/issues/54). Euring scheme: 01730.  We retrieve the values of field `Nummer` from `birds` related to this species:

```{r common_shelduck_nummers}
common_shellduck_nummer <-
  birds %>%
  filter(EuringCode == "01730") %>%
  pull(Nummer)
```

Number of birds to remove:

```{r n_common_shelduck}
length(common_shellduck_nummer)
```

We remove data from `birds`:

```{r remove_common_shelduck_from_birds}
birds <- birds %>%
  filter(EuringCode != "01730" | is.na(EuringCode))
```

and from `obs_and_acts`:

```{r remove_common_shelduck_from_obs_and_acts}
obs_and_acts <-
  obs_and_acts %>%
  filter(!KleurringNummer %in% common_shellduck_nummer | is.na(KleurringNummer))
```

## Remove rings `BGAH`, `BRAB` and `EAU`{#remove-three-rings}

The rings `BGAH`, `BRAB` and `EAU`should be removed, as explained in [# 89-issuesomment-481195862](https://github.com/inbo/cr-birding/issues/89#issuecomment-481195862) and [# 102-issuecomment-604939290](https://github.com/inbo/cr-birding/issues/102#issuecomment-604939290). No information about EURING code can be found:

```{r show_BGAH_BRAB}
birds %>%
  filter(Nummer %in% c("BGAH", "BRAB", "EAU"))
```

We also remove the ring `EAU` as it is not valid and the INBO experts agree on removing it (see [issue #102](https://github.com/inbo/cr-birding/issues/102)).

We remove these rings from `birds`:

```{r remove_BGAH_BRAB_EAU_from_birds}
birds <-
  birds %>%
  filter(!Nummer %in% c("BGAH", "BRAB", "EAU"))
```

Observations related to these three rings:

```{r show_obs_BGAH_BRAB}
obs_and_acts %>%
  filter(KleurringNummer %in% c("BGAH", "BRAB", "EAU"))
```

will be also removed:

```{r remove_obs_of_BGAH_BRAB}
obs_and_acts <-
  obs_and_acts %>%
  filter(!KleurringNummer %in% c("BGAH", "BRAB", "EAU"))
```

## Check date of `dood`

Observations with action `dood` should be the very last for each observation sequence containing this action. Exceptions:

```{r obs_after_death}
obs_after_death <-
  obs_and_acts %>%
  filter(dood == "dood") %>%
  select(KleurringNummer, Datum) %>%
  rename(dood_datum = Datum) %>%
  left_join(obs_and_acts, by = "KleurringNummer") %>%
  filter(dood_datum < Datum) %>%
  select(Nummer, KleurringNummer, Datum, dood_datum, acts, everything()) %>%
  arrange(KleurringNummer, Datum)
obs_after_death
```

If present, observations after `dood` are removed:

```{r remove_obs_after_death}
obs_and_acts <-
  obs_and_acts %>%
  filter(!Nummer %in% obs_after_death$Nummer)
```

Finally, we perform a last check: a ring should not be linked to two observations, one with action  `dood` and one with action `klgev`:

```{r check_dood_klgev}
obs_and_acts %>%
  filter(dood == "dood" | klgev == "klgev") %>%
  group_by(KleurringNummer) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
```

## Save preprocessed data

After data cleaning we save the data as TSVs in `.data/interim`:

```{r save_preprocessed_data}
write_tsv(users, path = here::here("data", "interim", "users.tsv"), na = "")
write_tsv(birds, path = here::here("data", "interim", "birds.tsv"), na = "")
write_tsv(
  obs_and_acts,
  path = here::here("data", "interim", "obs_and_actions.tsv"),
  na = ""
)
```

We save the tables containing action codes, colors and ring places in `./data/input` as it could be useful in following mapping steps:

```{r save_actions_meaning}
# Save action codes
write_tsv(
  actions_meaning,
  path = here::here("data", "input", "actions_meaning.tsv"),
  na = ""
)

# Save colors
write_tsv(
  color_table,
  path = here::here("data", "input", "color_table.tsv"),
  na = ""
)

# Save ring position
write_tsv(
  ring_position_table,
  path = here::here("data", "input", "ring_position_table.tsv"),
  na = ""
)
```
