==> src/2_users.Rmd <==
# Map user data

## Read temporary user data

We start from the user temporary data saved in TSV file `users.tsv` in folder `data\interim`:

```{r import_temp_users}
users <- read_tsv(here::here("data", "interim", "users.tsv"),
  col_types = cols(
    .default = col_character(),
    Nummer = col_double(),
    Wachtwoord = col_logical()
  )
)
```

Number of users:

```{r n_row_users}
nrow(users)
```

## Map user data

We map the original fields to SOVON fields denoted by prefix `sovon_`.

### E-mail

```{r e-mail}
users <-
  users %>%
  mutate(sovon_user_email = Email)
```

### First name

```{r first_name}
users <-
  users %>%
  mutate(sovon_user_first_name = Voornaam)
```

### Last name

```{r last_name}
users <-
  users %>%
  mutate(sovon_user_last_name = Familienaam)
```

### Address

```{r address}
users <-
  users %>%
  mutate(sovon_user_address = Adres)
```

### Place

```{r place}
users <-
  users %>%
  mutate(sovon_user_place = Gemeente)
```

### Postal code

```{r postal_code}
users <-
  users %>%
  mutate(sovon_user_postal_code = Postcode)
```

### Country

Countries present:

```{r countries_in_users}
users %>%
  distinct(LandCode)
```

```{r country}
users <-
  users %>%
  mutate(sovon_user_country = LandCode)
```

### User ID

User identifiers are provided by SOVON. `NA` is given.

```{r userID}
users <-
  users %>%
  mutate(sovon_user_id = NA)
```

### User reference

We use the unique ID in `Nummer`:

```{r user_ref}
users <-
  users %>%
  mutate(sovon_user_reference = Nummer)
```

### User language

This field is not present in `users`. We leave it empty:

```{r user_language}
users <-
  users %>%
  mutate(sovon_user_language = NA_character_)
```

## Save user data

Export the SOVON fields to `crbirding_users`:

```{r make_crbirding_users}
crbirding_users <-
  users %>%
  select(starts_with("sovon_"))
```

Remove prefix `sovon_`:

```{r remove prefix_sovon_users}
names(crbirding_users) <- str_remove(
  names(crbirding_users),
  pattern = "sovon_"
)
```

The desired order of columns in `crbirding_users`:

```{r cols_order_users}
cr_users_cols <- c(
  "user_id", "user_reference", "user_email", "user_first_name",
  "user_last_name", "user_address", "user_postal_code", "user_place",
  "user_country", "user_language", "user_role"
)
```

Fields still not mapped:

```{r users_fields_not_yet_mapped}
cr_users_cols[which(!cr_users_cols %in% names(crbirding_users))]
```

The field `user_role` cannot be filled at the moment: it will be mapped at the end of the next chapter.

Set column order:

```{r get_right_order_cols_users}
crbirding_users <-
  crbirding_users %>%
  select(cr_users_cols[cr_users_cols != "user_role"])
```

Preview data (e-mail, first and last names removed for privacy reasons):

```{r final_preview_users}
crbirding_users %>%
  select(-c(user_email, user_first_name, user_last_name)) %>%
  head(n = 10)
```

Save to csv file `crbirding_users.csv` in `./data/processed/` as asked by SOVON:

```{r write_user_data_txt}
crbirding_users %>%
  write_csv(
    path = here::here("data", "processed", "crbirding_users.csv"),
    na = ""
  )
```
