==> src/3_birds.Rmd <==
# Map ring data

## Read temporary ring data 

Import temporary bird data from `birds.tsv`:

```{r import_temp_birds}
birds <- read_tsv(
  here::here("data", "interim", "birds.tsv"),
  col_types = cols(
    .default = col_character()
  )
)
```

### Read observation data

Import temporary observation data from `obs_and_actions.tsv`:

```{r import_temp_obs_and_acts}
obs_and_acts <- read_tsv(
  here::here("data", "interim", "obs_and_actions.tsv"),
  col_types = cols(
    .default = col_character(),
    Nummer = col_double(),
    Datum = col_datetime(format = ""),
    WaarnemerNummer = col_double(),
    PlaatsLengtegraadGraden = col_double(),
    PlaatsLengtegraadMinuten = col_double(),
    PlaatsLengtegraadSeconden = col_double(),
    PlaatsBreedtegraadGraden = col_double(),
    PlaatsBreedtegraadMinuten = col_double(),
    PlaatsBreedtegraadSeconden = col_double(),
    BevestigingDatum = col_datetime(format = ""),
    AanmaakDatum = col_datetime(format = ""),
    WijzigDatum = col_datetime(format = "")
  )
)
```

### Read action data

Import action codes and relative meaning:

```{r import_actions_meaning_table}
actions_meaning <- read_tsv(here::here("data", "input", "actions_meaning.tsv"))
```

### Read ring position data

```{r import_ring_place_table}
ring_position_table <- read_tsv(
  here::here("data", "input", "ring_position_table.tsv")
)
```

### Read user data

Import user data as we have still to map the field `user_role`:

```{r import_temp_user_data}
crbirding_users <- read_csv(
  here::here("data", "processed", "crbirding_users.csv")
)
```

### Read color data

Import ring color codes and relative meaning:

```{r import_colors}
color_table <- read_tsv(here::here("data", "input", "color_table.tsv"))
```

## Map color ring data

### Extract action codes

Actions present in `obs_and_acts` :

```{r get_actions_in_obs_and_acts}
acts <- actions_meaning$Code
acts <- acts[acts %in% names(obs_and_acts)]
acts
```

### Bird reference & bird shorthand

In SOVON table `crbirding_birds` each row identifies a ring. By assigning a bird reference as unique integer to a bird they can link any color ring to the bird it belongs to.

```{r sovon_bird_reference}
birds <-
  birds %>%
  mutate(sovon_bird_reference = seq_len(nrow(birds))) %>%
  select(
    sovon_bird_reference, first_Nummer, NummerNieuw,
    NummerDesc, everything()
  )
head(birds)
```

Birds with lack of unicity of `sovon_bird_reference`:

```{r issues_shorthand}
birds %>%
  group_by(first_Nummer) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1) %>%
  select(
    sovon_bird_reference,
    first_Nummer,
    NummerNieuw,
    NummerDesc,
    Nummer
  )
```

In such cases we assign the lower `sovon_bird_reference`:

```{r solve_unicity_issues}
birds <-
  birds %>%
  group_by(first_Nummer) %>%
  mutate(sovon_bird_reference = min(sovon_bird_reference)) %>%
  ungroup()
```

Check:

```{r check_solution_unicity_shorthand}
birds %>%
  group_by(first_Nummer) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1) %>%
  select(
    sovon_bird_reference,
    first_Nummer,
    NummerNieuw,
    NummerDesc,
    Nummer
  )
```

We add column `sovon_bird_reference` to `obs_and_acts` as it will be needed to map the dates of ringing. The link is made by the columns `Nummer` (birds data) and  `KleurringNummer` (observation data):

```{r apply_bird_ref_to_obs_acts}
obs_and_acts <-
  obs_and_acts %>%
  left_join(birds %>% select(Nummer, sovon_bird_reference),
    by = c("KleurringNummer" = "Nummer")
  ) %>%
  select(sovon_bird_reference, everything())
```

Check whether there are observations without link to any bird:

```{r check_NA_bird_ref}
filter(obs_and_acts, is.na(sovon_bird_reference))
```

We will also add column `first_Nummer` from birds to observations as well:

```{r add_first_Nummer_to_obs_and_acts}
obs_and_acts <-
  obs_and_acts %>%
  left_join(
    birds %>%
      select(sovon_bird_reference, first_Nummer, Nummer),
    by = c("sovon_bird_reference",
      "KleurringNummer" = "Nummer"
    )
  )
```

The reconstruction of the series of color rings for each bird is made by using columns `first_Nummer` (very first color ring), `NummerNieuw` (very last color ring) and the information from spreadsheets of experts. We will collect the needed information in a new dataframe, called `crbirding_birds` which will grow up to become the end product containing the ring data to deliver to SOVON.

First step is to *gather* `first_Nummer` and `NummerNieuw` in new column `sovon_bird_shorthand`:

```{r backbone_mapping_shorthand}
crbirding_birds <-
  birds %>%
  select(
    sovon_bird_reference, first_Nummer,
    NummerNieuw, NummerDesc, sovon_bird_notes
  ) %>%
  pivot_longer(
    cols = c(first_Nummer, NummerNieuw),
    names_to = "col_nummer",
    values_to = "sovon_bird_shorthand"
  ) %>%
  select(-col_nummer) %>%
  select(sovon_bird_reference, sovon_bird_shorthand, NummerDesc, sovon_bird_notes)
```

We can also assign the color ring version with point (`sovon_bird_shorthand_pt`) by matching columns `sovon_bird_shorthand` and `NummerDesc` where possible. Note that `sovon_bird_shorthand_pt` may contain points, but for old rings it is equal to `sovon_bird_shorthand`:

```{r map_shorthand_with_point}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_shorthand_pt = ifelse(
    str_remove_all(
      string = NummerDesc,
      pattern = "\\."
    ) == sovon_bird_shorthand,
    NummerDesc,
    sovon_bird_shorthand
  )) %>%
  distinct() %>%
  select(sovon_bird_reference, sovon_bird_shorthand, sovon_bird_shorthand_pt, sovon_bird_notes)
## to map correctly MKAU (sovon_bird_reference: 4531)
## group_by(sovon_bird_reference, sovon_bird_shorthand) %>%
## filter(sovon_bird_shorthand ==
##          sovon_bird_shorthand[which.max(nchar(sovon_bird_shorthand))]) %>%
## ungroup()
```

Example of mapping: the bird with `sovon_bird_reference` 14 is associated to the following two rings:

```{r example_mapping_shorthand_birds}
birds %>%
  filter(sovon_bird_reference == 14) %>%
  select(sovon_bird_reference, first_Nummer, NummerNieuw, NummerDesc)
```

and will be mapped as follows:

```{r example_mapping_shorthand}
crbirding_birds %>%
  filter(sovon_bird_reference == 14) %>%
  select(sovon_bird_reference, sovon_bird_shorthand, sovon_bird_shorthand_pt)
```

### Bird ringing date

The date of applying a ring should be mapped as `sovon_bird_date_begin`. The date of applying the first ring can be found in data frame `obs_and_acts` in the column `Datum` for actions `rngkl` (code action of applying very first color ring):

```{r example_rngkl_actions}
obs_and_acts %>%
  filter(!is.na(rngkl)) %>%
  select(
    sovon_bird_reference, first_Nummer,
    Datum, rngkl
  ) %>%
  head(n = 10)
```

#### Date of applying first ring

We can add automatically the date of the very first ring for each bird (`sovon_bird_reference`), based on date of action code `rngkl`. In fact, there should be just one observation with action `rngkl` for each bird (`sovon_bird_reference`). Exceptions:

```{r exception_rule_one_bird_one_rngkl}
exceptions_one_bird_one_rngkl <-
  obs_and_acts %>%
  filter(!is.na(rngkl)) %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1) %>%
  filter(!is.na(rngkl)) %>%
  left_join(crbirding_birds %>%
    select(sovon_bird_reference, sovon_bird_shorthand),
  by = "sovon_bird_reference"
  ) %>%
  # distinct() %>%
  select(-n)
exceptions_one_bird_one_rngkl
```

In case exceptions are present, they should best be solved at database level. In any case, we will manage them later. 

Assign the date of first ringing:

```{r get_first_ring_date}
crbirding_birds <-
  obs_and_acts %>%
  filter(!is.na(rngkl)) %>%
  select(sovon_bird_reference, first_Nummer, Datum) %>%
  group_by(sovon_bird_reference, first_Nummer) %>%
  summarize(sovon_bird_date_begin = min(Datum)) %>%
  right_join(crbirding_birds,
    by = c("sovon_bird_reference",
      "first_Nummer" = "sovon_bird_shorthand"
    )
  ) %>%
  rename(sovon_bird_shorthand = first_Nummer) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_shorthand_pt,
    sovon_bird_date_begin,
    sovon_bird_notes
  ) %>%
  ungroup()
```

Some examples (birds with `sovon_bird_reference` 1, 14 and 4107):

```{r examples_date_ringing}
crbirding_birds %>%
  filter(sovon_bird_reference %in% c(1, 14, 4107)) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

Check date of birds in exceptions:

```{r check_date_of_exceptions_double_rngkl}
crbirding_birds %>%
  filter(sovon_bird_shorthand %in% exceptions_one_bird_one_rngkl$sovon_bird_shorthand)
```

And in case solve:

```{r solve_date_of_exceptions_double_rngkl, evaluate = FALSE}
# if (nrow(exceptions_one_bird_one_rngkl) > 0) {
#   crbirding_birds <-
#     crbirding_birds %>%
#     mutate(sovon_bird_date_begin = case_when(
#       sovon_bird_shorthand == "MKAU" ~ sovon_bird_date_begin,
#       sovon_bird_shorthand == "CZOZ" ~ as.POSIXct(NA_character_),
#       TRUE ~ sovon_bird_date_begin
#     ))
# }
```

Summary of rings with date (`sovon_bird_date_begin`):

```{r summary_n_rings_with_date}
crbirding_birds %>%
  mutate(date_is_present = !is.na(sovon_bird_date_begin)) %>%
  group_by(date_is_present) %>%
  count()
```

#### Date of applying last ring

The rings without date are the rings of birds ringed more than once. 

```{r summary_bird_with_more_rings}
birds_multiple_rings <-
  crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
birds_multiple_rings
```

Not only, the structure of our database limits to 2 the maximum number of rings linked to the same bird, as two are the columns containing such information (`Nummer` and `NummerNieuw`). Check:

```{r check_max_rings_is_2}
birds_two_rings <-
  crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n == 2)
all(birds_two_rings$sovon_bird_reference ==
  birds_multiple_rings$sovon_bird_reference)
```

Experts explained us that the very first and the very last ring are mapped in `birds`. By consulting the spreadsheet of the experts we will later fill the gap by adding _intermediate rings_.

We can try to retrieve the date of applying last ring based on observations/actions with code `vang` (caught at the nest) or `vangl` (caught otherwise). Birds ringed twice and linked to **one** `vang`/`vangl` action only:

```{r birds_two_rings_one_vang}
bird_one_vang <-
  obs_and_acts %>%
  filter(sovon_bird_reference %in%
    birds_multiple_rings$sovon_bird_reference &
    (!is.na(vang) | !is.na(vangl))) %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n == 1) %>%
  rename(sovon_bird_date_begin = Datum) %>%
  select(-n) %>%
  arrange(sovon_bird_reference)
bird_one_vang %>%
  select_if(function(x) any(!is.na(x)))
```

As you cannot change a ring to a bird without catching him, we can use these unique dates as date of ringing, `sovon_bird_date_begin`:

```{r set_date_ringing_bird_catched_once}
bird_one_vang_with_date <-
  crbirding_birds %>%
  filter(sovon_bird_reference %in% bird_one_vang$sovon_bird_reference &
    sovon_bird_reference %in% birds_two_rings$sovon_bird_reference) %>%
  group_by(sovon_bird_reference) %>%
  filter(is.na(sovon_bird_date_begin)) %>%
  select(-sovon_bird_date_begin) %>%
  ungroup() %>%
  left_join(bird_one_vang,
    by = c("sovon_bird_reference")
  ) %>%
  select(names(crbirding_birds))

crbirding_birds <-
  crbirding_birds %>%
  anti_join(bird_one_vang_with_date,
    by = c("sovon_bird_reference", "sovon_bird_shorthand")
  ) %>%
  bind_rows(bird_one_vang_with_date) %>%
  arrange(sovon_bird_reference)
```

Examples of date mapping (birds with `sovon_bird_reference` 11 and 14):

```{r examples_assign_date_second_ring_by_vang_date}
crbirding_birds %>%
  filter(sovon_bird_reference %in% c(11, 14))
```

Some birds have been catched twice:

```{r birds_two_rings_two_vang}
bird_two_vang <-
  obs_and_acts %>%
  filter(sovon_bird_reference %in% birds_two_rings$sovon_bird_reference &
    (!is.na(vang) | !is.na(vangl))) %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n == 2) %>%
  rename(sovon_bird_date_begin = Datum) %>%
  select(-n) %>%
  arrange(sovon_bird_reference)
bird_two_vang %>%
  select_if(function(x) any(!is.na(x)))
```

In the expert spreadsheet, we can see that the color ring is **typically** applied during the last `vang`/`vangl` action. We will assign the most recent date by default as first step, correcting the exceptions later:

```{r get_second_ring_via_vang}
bird_two_vang_get_date <-
  crbirding_birds %>%
  filter(sovon_bird_reference %in% bird_two_vang$sovon_bird_reference &
    sovon_bird_reference %in% birds_two_rings$sovon_bird_reference) %>%
  group_by(sovon_bird_reference) %>%
  filter(is.na(sovon_bird_date_begin)) %>%
  select(-sovon_bird_date_begin) %>%
  left_join(bird_two_vang %>%
    group_by(sovon_bird_reference) %>%
    filter(sovon_bird_date_begin == max(sovon_bird_date_begin)),
  by = c("sovon_bird_reference")
  ) %>%
  select_if(function(x) any(!is.na(x)))
bird_two_vang_get_date
```

By consulting the spreadsheet we can find that the following rings have been applied during the earliest `vang`/`vangl` action, so they are exceptions:

```{r correct_zexceptions_bird_date_begin_earliest_date}
bird_shorthand_exceptions <- c(
  "NGAP", "NGAX", "GVAR",
  "KPAZ", "KAAK", "KAAN"
)
bird_exceptions <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand %in% bird_shorthand_exceptions)
```

We set `sovon_bird_date_begin` equal to the date of the earliest `vang`/`vangl` action:

```{r set_date_to_earliest_vang}
bird_two_vang_get_date <-
  bird_two_vang %>%
  filter(sovon_bird_reference %in% bird_exceptions$sovon_bird_reference) %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin == min(sovon_bird_date_begin)) %>%
  ungroup() %>%
  left_join(bird_exceptions %>%
    select(-sovon_bird_date_begin), by = "sovon_bird_reference") %>%
  select(names(bird_two_vang_get_date)) %>%
  bind_rows(bird_two_vang_get_date %>%
    filter(!sovon_bird_shorthand %in%
      bird_exceptions$sovon_bird_shorthand)) %>%
  select(names(crbirding_birds))

bird_two_vang_get_date
```

Add this information to `crbirding_birds`:

```{r add_date_birds_with_two_vang}
crbirding_birds <-
  crbirding_birds %>%
  anti_join(bird_two_vang_get_date,
    by = c(
      "sovon_bird_reference",
      "sovon_bird_shorthand",
      "sovon_bird_shorthand_pt"
    )
  ) %>%
  bind_rows(bird_two_vang_get_date) %>%
  arrange(sovon_bird_reference)
crbirding_birds
```

Rings still without date:

```{r summary_date_begin_after}
crbirding_birds %>%
  mutate(date_is_present = !is.na(sovon_bird_date_begin)) %>%
  filter(!date_is_present)
```

Based on spreadsheet and observation data, we found that the dates of applying the rings FAAG, YCAF, FHOV and HSOW are linked to actions `ziek`, i.e. the ring has been applied while taking care of the birds. 

The date of applying FAAG:

```{r get_info_FAAG}
info_faag <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand == "FAAG") %>%
  mutate(
    sovon_bird_date_begin =
      obs_and_acts %>%
        filter(
          first_Nummer == "ALAU",
          !is.na(ziek)
        ) %>%
        ## get the second "ziek" action (2006-07-07)
        filter(Datum == max(Datum)) %>%
        pull(Datum)
  )
info_faag
```

The date of applying YCAF:

```{r get_info_YCAF}
info_ycaf <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand == "YCAF") %>%
  mutate(
    sovon_bird_date_begin =
      obs_and_acts %>%
        filter(
          first_Nummer == "PLAB",
          !is.na(ziek)
        ) %>%
        ## there is just one "ziek" action (2013-07-22)
        pull(Datum)
  )
info_ycaf
```

The date of applying FHOV:

```{r get_info_FHOV}
info_fhov <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand == "FHOV") %>%
  mutate(
    sovon_bird_date_begin =
      obs_and_acts %>%
        filter(
          first_Nummer == "SUAV",
          !is.na(ziek)
        ) %>%
        ## there is just one "ziek" action (2018-04-26)
        pull(Datum)
  )
info_fhov
```

The date of applying HSOW:

```{r get_info_HSOW}
info_hsow <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand == "HSOW") %>%
  mutate(
    sovon_bird_date_begin =
      obs_and_acts %>%
        filter(
          first_Nummer == "CPAG",
          !is.na(ziek)
        ) %>%
        ## there is just one "ziek" action (2018-04-26)
        pull(Datum)
  )
info_hsow
```

Based on notes (column `Opmerking` in `obs_and_acts`) we understand that the color ring BUAH has been placed while applying the metal ring, so it is linked to action `rngme`.

The date of applying BUAH:

```{r get_info_buah}
info_buah <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand == "BUAH") %>%
  mutate(
    sovon_bird_date_begin =
      obs_and_acts %>%
        filter(
          first_Nummer == "ZBZZ",
          !is.na(rngme)
        ) %>%
        pull(Datum)
  )
info_buah
```

Add this dates to `crbirding_birds`:

```{r add_to_crbirding_birds}
crbirding_birds <-
  bind_rows(
    crbirding_birds %>%
      filter(!sovon_bird_shorthand %in% c("FAAG",
                                          "YCAF",
                                          "FHOV",
                                          "HSOW",
                                          "BUAH")),
    info_faag,
    info_ycaf,
    info_fhov,
    info_hsow,
    info_buah
  ) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

All rings in `crbirding_birds` have a date:

```{r check_presence_date}
crbirding_birds %>%
  filter(!is.na(sovon_bird_date_begin)) %>%
  nrow() == nrow(crbirding_birds)
```

#### Add intermediate color rings

Based on spreadsheet of experts, we have to add new rings, as some birds have been ringed more than twice, so the information contained in `first_Nummer` and `NummerNieuw` is not complete, as they map the very first and the very last ring. The intermediate rings are the following:

```{r intermediate_rings_df}
intermediate_rings <- tibble(
  first_Nummer = c(
    "E633",
    "DJAB",
    "BWAD",
    "KLAT",
    "PR3",
    "TY2",
    "RTO",
    "E099",
    "MKAU",
    "ASAH",
    "ADAF",
    "KMAV",
    "KCAV",
    "BCAV"
  ),
  intermediate_ring = c(
    "BUAG",
    "TRAP",
    "GTAS",
    "UPAB",
    "DHAZ",
    "GMAJ",
    "DGAH",
    "GHAT",
    "ZVAU",
    "HWAX",
    "DJAU",
    "UMAR",
    "RTAK",
    "LMAA"
  ),
  sovon_bird_date_begin = as.POSIXct(c(
    "2000-05-08", ## BUAG
    "2012-05-25", ## TRAP
    "2007-06-01", ## GTAS
    "2012-05-29", ## UPAB
    "2006-05-26", ## DHAZ
    "2007-06-06", ## GMAJ
    "2006-05-17", ## DGAH
    "2007-06-18", ## GHAT
    "2014-05-23", ## ZVAU
    "2009-05-18", ## HWAX
    "2006-05-26", ## DJAU
    "2012-05-31", ## UMAR
    "2016-05-26", ## RTAK
    "2009-06-01" ## LMAA
  ), tz = "UTC")
)
intermediate_rings
```

Retrieve `sovon_bird_reference`:

```{r retrieve_bird_ref_intermediate_rings}
intermediate_rings <-
  intermediate_rings %>%
  left_join(crbirding_birds %>%
    select(sovon_bird_reference, sovon_bird_shorthand, sovon_bird_notes),
  by = c("first_Nummer" = "sovon_bird_shorthand")
  ) %>%
  select(
    sovon_bird_reference, first_Nummer, intermediate_ring,
    sovon_bird_date_begin, sovon_bird_notes
  )
intermediate_rings
```

Check whether all dates are linked to a valid observation and check whether they are linked to `vang/vangl` actions:

```{r get_actins_intermediate_rings}
intermediate_rings %>%
  left_join(obs_and_acts %>%
    select(
      sovon_bird_reference,
      Datum,
      acts,
      first_Nummer
    ),
  by = c("sovon_bird_reference",
    "sovon_bird_date_begin" = "Datum",
    "first_Nummer"
  )
  ) %>%
  filter(!is.na(vang) | !is.na(vangl)) %>%
  select_if(function(x) any(!is.na(x)))
```

We can then add the intermediate rings to `crbirding_birds`:

```{r add_intermediate_ring}
crbirding_birds <-
  crbirding_birds %>%
  bind_rows(intermediate_rings %>%
    select(-first_Nummer) %>%
    rename(sovon_bird_shorthand = intermediate_ring) %>%
    mutate(sovon_bird_shorthand_pt = sovon_bird_shorthand)) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

These are all mapped triplet ring sequences:

```{r triplet_ring_sequences}
bird_reference_triplets <-
  crbirding_birds %>%
  filter(sovon_bird_shorthand %in% c(intermediate_rings$intermediate_ring)) %>%
  distinct(sovon_bird_reference) %>%
  pull()
crbirding_birds %>%
  filter(sovon_bird_reference %in% bird_reference_triplets)
```

At this point all rings should have a `sovon_bird_date_begin`:

```{r final_check_all_rings_with_date}
crbirding_birds %>%
  filter(!is.na(sovon_bird_date_begin)) %>%
  nrow() == nrow(crbirding_birds)
```

### Bird ringing end date

For birds ringed more than once, we can assign an end date for the changed rings. This information will be stored in field `sovon_bird_date_end`. The end date is equal to the date of applying the new ring

```{r assign_bird_date_end}
assign_end_date <- function(data) {
  if (nrow(data) > 1) {
    return(c(
      as_date(data$sovon_bird_date_begin[2:nrow(data)], tz = "UTC"),
      as_date(NA, tz = "UTC")
    ))
  } else {
    return(as_date(NA, tz = "UTC"))
  }
}

crbirding_birds <-
  crbirding_birds %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin) %>%
  group_by(sovon_bird_reference) %>%
  nest() %>%
  mutate(sovon_bird_date_end = map(data, assign_end_date)) %>%
  unnest(cols = c(data, sovon_bird_date_end)) %>%
  mutate(sovon_bird_date_end = as.POSIXct(sovon_bird_date_end)) %>%
  select(
    sovon_bird_reference, sovon_bird_shorthand, sovon_bird_shorthand_pt,
    sovon_bird_date_begin, sovon_bird_date_end, sovon_bird_notes
  ) %>%
  ungroup()
```

As example, we show the chronology of color rings for birds ringed thrice:

```{r examples_bird_date_end}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n == 3) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end
  ) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

We add end date to the very last ring of dead birds by using the link to observations with action `dood`:

```{r add_end_date_to_dead_birds}
crbirding_birds <-
  obs_and_acts %>%
  filter(dood == "dood") %>%
  select(sovon_bird_reference, Datum) %>%
  rename(bird_dood_datum = Datum) %>%
  right_join(crbirding_birds,
    by = c("sovon_bird_reference")
  ) %>%
  group_by(sovon_bird_reference) %>%
  mutate(sovon_bird_date_end = case_when(
    is.na(sovon_bird_date_end) &
      sovon_bird_date_begin == max(sovon_bird_date_begin) &
      bird_dood_datum >= sovon_bird_date_begin ~ bird_dood_datum,
    TRUE ~ sovon_bird_date_end
  )) %>%
  select(-bird_dood_datum)
```

Effects of mapping:

```{r show_death_date_mapping}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin == max(sovon_bird_date_begin) &
    !is.na(sovon_bird_date_end)) %>%
  select(sovon_bird_reference) %>%
  ungroup() %>%
  left_join(crbirding_birds, by = "sovon_bird_reference") %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    everything()
  )
```

Number of rings still in use, i.e. no end date: 

```{r rings_in_use}
crbirding_birds %>%
  filter(is.na(sovon_bird_date_end)) %>%
  nrow()
```

### Add rows for metal rings applied before color rings

Sometimes metal rings have been applied **before** color rings. These cases should be added to `crbirding_birds` as new rows. We can detect these situations by comparing the date of actions `rngme` and `rngkl` for each bird.

```{r compare_rngme_rngkl}
rngme_before_rngkl <-
  obs_and_acts %>%
  filter(!is.na(rngme) & is.na(rngkl)) %>%
  select(sovon_bird_reference, Datum) %>%
  rename(Datum_rngme = Datum) %>%
  left_join(obs_and_acts %>%
    filter(!is.na(rngkl)) %>%
    select(sovon_bird_reference, Datum) %>%
    rename(Datum_rngkl = Datum),
  by = "sovon_bird_reference"
  ) %>%
  filter(Datum_rngme < Datum_rngkl) %>%
  rename(
    sovon_bird_date_begin = Datum_rngme,
    sovon_bird_date_end = Datum_rngkl
  ) %>%
  mutate(
    sovon_bird_shorthand = NA_character_,
    sovon_bird_shorthand_pt = NA_character_,
    sovon_bird_notes = "sovon_bird_shorthand not available."
  ) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_shorthand_pt,
    everything()
  )
rngme_before_rngkl
```

We add these rows to `crbirding_birds`:

```{r add_rngme_before_rngklt_to_crbirding_birds}
crbirding_birds <-
  crbirding_birds %>%
  bind_rows(rngme_before_rngkl) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

A preview:

```{r preview_rngme_before_rngkl}
crbirding_birds %>%
  filter(sovon_bird_reference %in%
    (rngme_before_rngkl %>%
      pull(sovon_bird_reference))) %>%
  head(n = 50)
```

### Ring number

The _ring number_, or _metal ring number_, can be found in field `MetaalringNummer`, which contains the **most recent** metal ring number. Recoverying the complete chronology of metal rings for each bird is almost impossible as it is:

1. not complete
2. included in texutal description (field `Opmerking` of `tblWaarneming`)

However, this is not considered a real problem by INBO and SOVON experts as both INBO and SOVON databases focus on color rings. See [issue 34](https://github.com/inbo/sovon/issues/34) for more details. Just as example, we can show the textual notes of observations linked to action `vang` and containing 6 or more consecutive numbers in the notes:

```{r example_notes_changes_metalring}
obs_and_acts %>%
  filter(str_detect(Opmerking, pattern = "[0-9]{6,}")) %>%
  filter(!is.na(vang)) %>%
  select(Opmerking)
```

A ring number should consist of up to ten **alphanumeric** characters. At page 7 of the [EURING Exchange Code 2000+ document](https://euring.org/files/documents/E2000PLUSExchangeCodev1161.pdf) we read the following instructions:

> Where the ring number consist of fewer than ten numbers/letters, the number is padded with dots. These dots are always inserted to the left of the rightmost
row of numbers.

Ring number anomalies:

```{r anomalies_ring_number}
birds %>%
  filter(!str_detect(MetaalringNummer, pattern = regex("\\d+"))) %>%
  group_by(MetaalringNummer) %>%
  count()
```

Value `none` means _not metal-ringed bird_, while `onbekend` means that information is no more available. SOVON chooses to leave column `sovon_bird_ring_number` empty for both cases. However, we will add a note to `sovon_bird_notes` to still maintain a reference to this slight difference. Value `?` is equivalent to `onbekend` while `H????????` and `Lxxxxxx` are mapped as `H--------` and `L------` respectively.

```{r solve_anomalies_bird_ring_number}
birds <-
  birds %>%
  mutate(sovon_bird_ring_number = recode(
    MetaalringNummer,
    "?" = NA_character_,
    "onbekend" = NA_character_,
    "none" = NA_character_,
    "H????????" = "H--------",
    "Lxxxxxx" = "L------"
  ))
```

We separate the birds with special mapping values by all the others:

```{r hold_special_values}
special_values <- c("H--------", "L------")
birds_special_values <-
  birds %>%
  filter(sovon_bird_ring_number %in% special_values)
birds_others <-
  birds %>%
  filter(!sovon_bird_ring_number %in% special_values)
```

Some ring numbers contain asterisks:

```{r show_asterisks}
birds_others %>%
  filter(str_detect(sovon_bird_ring_number, "\\*")) %>%
  select(sovon_bird_ring_number)
```

We remove the asterisks:

```{r remove_*_in_ring_number}
birds_others <-
  birds_others %>%
  mutate(sovon_bird_ring_number = str_remove_all(sovon_bird_ring_number, "\\*"))
```

Preview:

```{r preview_remove_asterisks}
birds_others %>%
  select(MetaalringNummer, sovon_bird_ring_number) %>%
  filter(str_detect(MetaalringNummer, "\\*")) 
```

Add points `.` where needed:

```{r add_points_in_ring_number}
birds_others <-
  birds_others %>%
  # calculate the number of dots needed
  mutate(ndots = 10 - str_length(sovon_bird_ring_number)) %>%
  # created a string with the needed dots
  mutate(dots = str_dup(".", ndots)) %>%
  # concatenate the first uppercase letter, the dots and the digits together
  mutate(sovon_bird_ring_number = str_c(
    str_extract(sovon_bird_ring_number, "^[:upper:]+"),
    dots,
    str_extract(sovon_bird_ring_number, "[:digit:]+$"))) %>%
  # remove auxiliary columns after use
  select(-c(ndots, dots))
```

Merge the two data frames together:

```{r merge_back_together}
birds <- bind_rows(birds_others, birds_special_values)
```

Preivew of the effects the mapping:

```{r show_effect_mapping_example}
birds %>%
  distinct(MetaalringNummer, sovon_bird_ring_number) %>%
  head(n = 200)
```

We have to be sure that no multiple metal rings are assigned to the same bird (`sovon_bird_reference`) as this is not allowed by our database. Exceptions:

```{r multiple_metal_rings_bird_reference}
birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
```

Later we will add value `sovon_bird_ring_number` to `crbirding_birds` where metal ring is present.

### Initialize `sovon_bird_rings_changed` column for very first ring

In order to better follow the mapping of metal AND color rings, SOVON suggested to add a column called `sovon_bird_rings_changed`, as shown in this [GitHub comment](https://github.com/inbo/cr-birding/issues/47#issuecomment-469256635).

We initialize the column `sovon_bird_rings_changed` for the **very first** ringing by assigning:

1. `sovon_bird_rings_changed` = 0 if `rngme` action only
2. `sovon_bird_rings_changed` = 1 if `rngkl` action only
3. `sovon_bird_rings_changed` = 2 if `rngkl` and `rngme`

If very first `bird_shorthand` is empty, it means that the bird has been first ringed with a metal ring, `sovon_bird_rings_changed` = 0:

```{r initialize bird_rings_changed_0}
crbirding_birds <-
  crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  mutate(sovon_bird_rings_changed = if_else(
    sovon_bird_date_begin == min(sovon_bird_date_begin) &
      is.na(sovon_bird_shorthand),
    0, NA_real_
  )) %>%
  ungroup()
```

We proceed now by mapping values 1 (action `rngkl` only) and 2 (action `rngkl` and `rngme` together):

```{r bird_rings_changed_1_or_2}
crbirding_birds <-
  crbirding_birds %>%
  left_join(obs_and_acts %>%
    filter(rngkl == "rngkl") %>%
    mutate(n_rings = if_else(is.na(rngme), 1, 2)) %>%
    select(sovon_bird_reference, n_rings, Datum),
  by = "sovon_bird_reference"
  ) %>%
  group_by(sovon_bird_reference) %>%
  mutate(sovon_bird_rings_changed = if_else(
    sovon_bird_date_begin == min(sovon_bird_date_begin) &
      !is.na(sovon_bird_shorthand),
    n_rings,
    sovon_bird_rings_changed
  )) %>%
  select(-c(n_rings, Datum)) %>%
  ungroup()
```

Preview `sovon_bird_rings_changed` = 1:

```{r preview_bird_rings_changed_1}
crbirding_birds %>%
  filter(sovon_bird_rings_changed == 1) %>%
  head(n = 10)
```

Preview `sovon_bird_rings_changed` = 2:

```{r preview_bird_rings_changed_2}
crbirding_birds %>%
  filter(sovon_bird_rings_changed == 2) %>%
  head(n = 10)
```

### Add help field `metal_ring_missing`

We add also a help column, `metal_ring_missing`: it will help us later to fill the field `sovon_ring_number` only where metal ring is present.

We initialize it as `FALSE` except for very first metal rings only (`sovon_bird_rings_changed` = 1):

```{r add_metal_ring_missing}
crbirding_birds <-
  crbirding_birds %>%
  mutate(metal_ring_missing = if_else(
    sovon_bird_rings_changed != 1 | is.na(sovon_bird_rings_changed),
    FALSE,
    TRUE
  ))
```

Preview `metal_ring_missing` = `FALSE`:

```{r preview_me_ring_not_missing}
crbirding_birds %>%
  filter(metal_ring_missing == FALSE) %>%
  head(n = 50)
```

Preview `metal_ring_missing` = `TRUE`:

```{r preview_me_ring_missing}
crbirding_birds %>%
  filter(metal_ring_missing == TRUE) %>%
  head(n = 50)
```

### Add rows for first metal ring applied after color rings

Sometimes metal rings have been applied **after** color rings. We can detect these situations by using `sovon_bird_rings_changed` value 1 (`rngkl` only) to identify the birds, observations with action `rngme` to get date of applying metal rings:

```{r rngkl_before_rngme}
rngme_after_rngkl <-
  crbirding_birds %>%
  filter(sovon_bird_rings_changed == 1) %>%
  select(sovon_bird_reference) %>%
  left_join(obs_and_acts %>%
    filter(rngme == "rngme") %>%
    select(sovon_bird_reference, Datum),
  by = "sovon_bird_reference"
  ) %>%
  select(sovon_bird_reference, Datum) %>%
  rename(sovon_bird_date_begin = Datum)
rngme_after_rngkl
```

If `sovon_bird_reference` and `sovon_bird_date_begin` are already present in `crbirding_birds` means that applying a metal ring occurred while changing color ring during a `vang` action: no rows should be added then, but `sovon_bird_rings_changed` = 2:

```{r set_bird_rings_changed_2_rngme}
crbirding_birds <-
  crbirding_birds %>%
  left_join(rngme_after_rngkl %>%
    rename(rngme_date = sovon_bird_date_begin),
  by = "sovon_bird_reference"
  ) %>%
  mutate(sovon_bird_rings_changed = case_when(
    sovon_bird_date_begin == rngme_date ~ 2,
    TRUE ~ sovon_bird_rings_changed
  )) %>%
  select(-rngme_date)
```

Show changes:

```{r show_rngme_while_changing_colorring}
rngme_after_rngkl %>%
  inner_join(crbirding_birds,
    by = c("sovon_bird_reference", "sovon_bird_date_begin")
  ) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    sovon_bird_rings_changed,
    metal_ring_missing,
    everything()
  )
```

On the other side, if `sovon_bird_reference` and `sovon_bird_date_begin` are not present in `crbirding_birds`, we add them as new rows where `sovon_bird_rings_changed` = 0 and `metal_ring_missing` is `FALSE`.

Rows to add:

```{r rows_rngme_to_add_crbirding_birds}
rngme_without_change_rngkl <-
  rngme_after_rngkl %>%
  anti_join(crbirding_birds, by = c("sovon_bird_reference", "sovon_bird_date_begin")) %>%
  rename(rngme_date = sovon_bird_date_begin) %>%
  right_join(crbirding_birds, by = "sovon_bird_reference") %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin < rngme_date) %>%
  filter(sovon_bird_date_begin == max(sovon_bird_date_begin)) %>%
  mutate(
    sovon_bird_rings_changed = 0,
    metal_ring_missing = FALSE,
    sovon_bird_date_begin = rngme_date
  ) %>%
  select(names(crbirding_birds)) %>%
  ungroup()
rngme_without_change_rngkl
```

Add these new rows:

```{r add_new_rings_rngme}
crbirding_birds <-
  crbirding_birds %>%
  bind_rows(rngme_without_change_rngkl) %>%
  arrange(
    sovon_bird_reference,
    sovon_bird_date_begin,
    sovon_bird_date_end
  ) %>%
  group_by(sovon_bird_reference, sovon_bird_date_end) %>%
  mutate(sovon_bird_date_end_updated = case_when(
    sovon_bird_date_begin == min(sovon_bird_date_begin) &
      sovon_bird_date_begin != max(sovon_bird_date_begin) ~ max(sovon_bird_date_begin),
    TRUE ~ sovon_bird_date_end
  )) %>%
  ungroup() %>%
  mutate(sovon_bird_date_end = sovon_bird_date_end_updated) %>%
  select(-sovon_bird_date_end_updated)
```

Show changes:

```{r show_changes_rngme_without_change_rngkl}
crbirding_birds %>%
  filter(sovon_bird_reference %in% c(rngme_without_change_rngkl$sovon_bird_reference))
```

For these birds, we can then assign value 1 to `sovon_bird_rings_changed` while applying color ring as `rngkl` occurs after `rngme`:

```{r assign_sovon_bird_rings_changed_1rngkl_after_rngme}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_rings_changed = case_when(
    sovon_bird_reference %in% rngme_before_rngkl$sovon_bird_reference &
      !is.na(sovon_bird_shorthand) ~ 1,
    TRUE ~ sovon_bird_rings_changed
  ))
```

Show changes in color/metal ring  for these birds:

```{r show_changes_rngme_before_rngkl}
crbirding_birds %>%
  filter(sovon_bird_reference %in% rngme_before_rngkl$sovon_bird_reference)
```

### Add rows for loosing color or metal rings: `klweg`,  `meweg`

We have observations linked to actions `klweg` and `meweg`. These actions mean that the bird has been observed without the color ring or metal ring respectively.

Number of observations with action `klweg`:

```{r show_klweg}
obs_and_acts %>%
  filter(klweg == "klweg") %>%
  nrow()
```

Type of action combinations where `klweg` occurs:

```{r combinations_klweg}
obs_and_acts %>%
  filter(klweg == "klweg") %>%
  select(acts) %>%
  distinct() %>%
  select_if(~ sum(!is.na(.)) > 0)
```

Number of observations with action `meweg`:

```{r show_meweg}
obs_and_acts %>%
  filter(meweg == "meweg") %>%
  nrow()
```

Type of action combinations where `meweg` occurs:

```{r combinations_meweg}
obs_and_acts %>%
  filter(meweg == "meweg") %>%
  select(acts) %>%
  distinct() %>%
  select_if(~ sum(!is.na(.)) > 0)
```

Are there birds linked to both actions `klweg` and `meweg`?

```{r bird_ref_klweg_meweg}
bird_ref_meweg <-
  obs_and_acts %>%
  filter(meweg == "meweg") %>%
  pull(sovon_bird_reference)
bird_ref_klweg <-
  (obs_and_acts %>%
    filter(klweg == "klweg") %>%
    pull(sovon_bird_reference))
bird_ref_meweg[which(bird_ref_meweg %in% bird_ref_klweg)]
```

Are there more than one `meweg` for bird?

```{r birds_meweg}
bird_ref_multiple_meweg <-
  obs_and_acts %>%
  filter(meweg == "meweg") %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
bird_ref_multiple_meweg
```

Are there more than one `klweg` for bird?

```{r birds_klweg}
bird_ref_multiple_klweg <-
  obs_and_acts %>%
  filter(klweg == "klweg") %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
bird_ref_multiple_klweg
```

If `meweg`/`klweg` occurs not in combination with `dood` or `vang`/ `vangl`, then a new row should be added to `crbirding_birds` as the ring situation changed. 

We create first new rows based on action `meweg`. We set `sovon_bird_rings_changed` = 0 and `metal_ring_missing` = `TRUE`:

```{r show_rings_to_add_meweg}
rings_to_add_meweg <-
  obs_and_acts %>%
  filter(meweg == "meweg" & is.na(dood) & is.na(vang) & is.na(vangl)) %>%
  select(sovon_bird_reference, Datum) %>%
  left_join(crbirding_birds, by = "sovon_bird_reference") %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin < Datum) %>%
  filter(sovon_bird_date_begin == max(sovon_bird_date_begin)) %>%
  mutate(sovon_bird_date_begin = Datum) %>%
  select(-Datum) %>%
  mutate(
    sovon_bird_rings_changed = 0,
    metal_ring_missing = TRUE
  ) %>%
  ungroup() %>%
  arrange(sovon_bird_reference)
rings_to_add_meweg
```

We add them, paying attention to change `bird_date_end` of previous rings:

```{r add_rings_meweg}
crbirding_birds <-
  crbirding_birds %>%
  bind_rows(rings_to_add_meweg) %>%
  arrange(
    sovon_bird_reference,
    sovon_bird_date_begin,
    sovon_bird_date_end
  ) %>%
  group_by(sovon_bird_reference, sovon_bird_date_end) %>%
  mutate(sovon_bird_date_end_updated = case_when(
    sovon_bird_date_begin == min(sovon_bird_date_begin) &
      sovon_bird_date_begin != max(sovon_bird_date_begin) ~
    max(sovon_bird_date_begin),
    TRUE ~ sovon_bird_date_end
  )) %>%
  ungroup() %>%
  mutate(sovon_bird_date_end = sovon_bird_date_end_updated) %>%
  select(-sovon_bird_date_end_updated)
```

and setting `sovon_bird_rings_changed` = 2 for the next ring, if exists:

```{r set_sovon_bird_rings_changed_2_for_next_ring}
rings_to_update <-
  rings_to_add_meweg %>%
  filter(!is.na(sovon_bird_date_end)) %>%
  left_join(
    obs_and_acts %>%
      filter(dood == "dood") %>%
      select(sovon_bird_reference, Datum) %>%
      rename(bird_dood_datum = Datum),
    by = "sovon_bird_reference"
  ) %>%
  filter(is.na(bird_dood_datum)) %>%
  select(-bird_dood_datum) %>%
  select(sovon_bird_reference, sovon_bird_date_end) %>%
  left_join(crbirding_birds %>%
    select(-sovon_bird_date_end),
  by = c("sovon_bird_reference",
    "sovon_bird_date_end" = "sovon_bird_date_begin"
  )
  ) %>%
  rename(sovon_bird_date_begin = sovon_bird_date_end) %>%
  mutate(
    sovon_bird_rings_changed = 2,
    metal_ring_missing = FALSE
  ) %>%
  left_join(crbirding_birds %>%
    select(-c(
      sovon_bird_rings_changed,
      metal_ring_missing
    )),
  by = c(
    "sovon_bird_reference",
    "sovon_bird_shorthand",
    "sovon_bird_shorthand_pt",
    "sovon_bird_date_begin",
    "sovon_bird_notes"
  )
  ) %>%
  select(names(crbirding_birds))
crbirding_birds <-
  crbirding_birds %>%
  anti_join(rings_to_update,
    by = c(
      "sovon_bird_reference",
      "sovon_bird_shorthand",
      "sovon_bird_shorthand_pt"
    )
  ) %>%
  bind_rows(rings_to_update) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

Show changes:

```{r show_changes_after_add_meweg}
crbirding_birds %>%
  filter(sovon_bird_reference %in%
    rings_to_add_meweg$sovon_bird_reference) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    sovon_bird_rings_changed,
    metal_ring_missing,
    everything()
  )
```

We create now new rows based on action `klweg`. We set `sovon_bird_rings_changed` = 1, `metal_ring_missing` = FALSE and we leave `bird_shorthand` empty:

```{r show_rings_to_add_klweg}
rings_to_add_klweg <-
  obs_and_acts %>%
  filter(klweg == "klweg" & is.na(dood) & is.na(vang) & is.na(vangl)) %>%
  select(sovon_bird_reference, Datum) %>%
  left_join(crbirding_birds, by = "sovon_bird_reference") %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin < Datum) %>%
  filter(sovon_bird_date_begin == max(sovon_bird_date_begin)) %>%
  mutate(sovon_bird_date_begin = Datum) %>%
  select(-Datum) %>%
  mutate(
    sovon_bird_rings_changed = 1,
    metal_ring_missing = FALSE,
    sovon_bird_shorthand = NA_character_,
    sovon_bird_shorthand_pt = NA_character_
  ) %>%
  ungroup() %>%
  arrange(sovon_bird_reference)
rings_to_add_klweg
```

We add then these new rows, paying attention to change `bird_date_end` of previous rings:

```{r add_rings_klweg}
crbirding_birds <-
  crbirding_birds %>%
  bind_rows(rings_to_add_klweg) %>%
  arrange(
    sovon_bird_reference,
    sovon_bird_date_begin,
    sovon_bird_date_end
  ) %>%
  group_by(sovon_bird_reference, sovon_bird_date_end) %>%
  mutate(sovon_bird_date_end_updated = case_when(
    sovon_bird_date_begin == min(sovon_bird_date_begin) &
      sovon_bird_date_begin != max(sovon_bird_date_begin)
    ~ max(sovon_bird_date_begin),
    TRUE ~ sovon_bird_date_end
  )) %>%
  ungroup() %>%
  mutate(sovon_bird_date_end = sovon_bird_date_end_updated) %>%
  select(-sovon_bird_date_end_updated)
```

Show changes:

```{r show_changes_klweg}
crbirding_birds %>%
  filter(sovon_bird_reference %in%
    rings_to_add_klweg$sovon_bird_reference) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    sovon_bird_rings_changed,
    metal_ring_missing,
    everything()
  )
```

### Add rows for found color or metal rings: `klgev`,  `megev`

We have observations linked to actions `klgev` and `megev`. These actions mean that a color/metal ring has been found. This means the bird is not wearing it anymore: we proceed as in previous section by adding a new row in `crbirding_birds`.

Number of observations with action `klgev`:

```{r show_klgev}
obs_and_acts %>%
  filter(klgev == "klgev") %>%
  nrow()
```

Type of action combinations where `klgev` occurs:

```{r combinations_klgev}
obs_and_acts %>%
  filter(klgev == "klgev") %>%
  select(acts) %>%
  distinct() %>%
  select_if(~ sum(!is.na(.)) > 0)
```

Number of observations with action `megev`:

```{r show_megev}
if ("megev" %in% acts) {
  obs_and_acts %>%
    filter(megev == "megev") %>%
    nrow()
} else {
  message("0")
}
```

Type of action combinations where `megev` occurs:

```{r combinations_megev}
if ("megev" %in% acts) {
  obs_and_acts %>%
    filter(megev == "megev") %>%
    select(acts) %>%
    distinct() %>%
    select_if(~ sum(!is.na(.)) > 0)
} else {
  message("'megev' is not present.")
}
```

Are there observations of birds linked to both actions `klgev` and `megev`?

```{r bird_ref_klgev_megev}
bird_ref_klgev <-
  (obs_and_acts %>%
    filter(klgev == "klgev") %>%
    pull(sovon_bird_reference))
if ("megev" %in% acts) {
  bird_ref_megev <-
    obs_and_acts %>%
    filter(megev == "megev") %>%
    pull(sovon_bird_reference)
  bird_ref_meweg[which(bird_ref_megev %in% bird_ref_klgev)]
} else {
  message("'megev' is not present.")
}
```

Are there observations of birds linked to action `klgev` and `klweg`?

```{r bird_ref_klgev_klweg}
bird_ref_klgev[which(bird_ref_klgev %in% bird_ref_klweg)]
```

Are there observations of birds linked to action `klgev` and `meweg`?

```{r bird_ref_klgev_meweg}
bird_ref_klgev[which(bird_ref_klgev %in% bird_ref_meweg)]
```

Are there more than one `megev` for bird?

```{r birds_megev}
if ("megev" %in% acts) {
  bird_ref_multiple_megev <-
    obs_and_acts %>%
    filter(megev == "megev") %>%
    group_by(sovon_bird_reference) %>%
    add_tally() %>%
    ungroup() %>%
    filter(n > 1)
  bird_ref_multiple_megev
} else {
  message("'megev' is not present.")
}
```

Are there more than one `klgev` for bird?

```{r birds_klgev}
bird_ref_multiple_klgev <-
  obs_and_acts %>%
  filter(klgev == "klgev") %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 1)
bird_ref_multiple_klgev
```

We add new rows to `crbirding_birds` as the ring situation changed at each of these observations/actions. 

We create first new rows based on action `megev`. We set `sovon_bird_rings_changed` = 0 and `metal_ring_missing` = `TRUE`:

```{r show_rings_to_add_megev}
if ("megev" %in% acts) {
  rings_to_add_megev <-
    obs_and_acts %>%
    filter(megev == "megev") %>%
    select(sovon_bird_reference, Datum) %>%
    left_join(crbirding_birds, by = "sovon_bird_reference") %>%
    group_by(sovon_bird_reference) %>%
    filter(sovon_bird_date_begin < Datum) %>%
    filter(sovon_bird_date_begin == max(sovon_bird_date_begin)) %>%
    mutate(sovon_bird_date_begin = Datum) %>%
    select(-Datum) %>%
    mutate(
      sovon_bird_rings_changed = 0,
      metal_ring_missing = TRUE
    ) %>%
    ungroup() %>%
    arrange(sovon_bird_reference)
  rings_to_add_megev
} else {
  message("'megev' is not present.")
}
```

We add them, paying attention to change `bird_date_end` of previous rings:

```{r add_rings_megev}
if ("megev" %in% acts) {
  crbirding_birds <-
    crbirding_birds %>%
    bind_rows(rings_to_add_megev) %>%
    arrange(
      sovon_bird_reference,
      sovon_bird_date_begin,
      sovon_bird_date_end
    ) %>%
    group_by(sovon_bird_reference, sovon_bird_date_end) %>%
    mutate(sovon_bird_date_end_updated = case_when(
      sovon_bird_date_begin == min(sovon_bird_date_begin) &
        sovon_bird_date_begin != max(sovon_bird_date_begin) ~
      max(sovon_bird_date_begin),
      TRUE ~ sovon_bird_date_end
    )) %>%
    ungroup() %>%
    mutate(sovon_bird_date_end = sovon_bird_date_end_updated) %>%
    select(-sovon_bird_date_end_updated)
} else {
  message("'megev' is not present.")
}
```

Show changes:

```{r show_changes_add_megev}
if ("megev" %in% acts) {
  crbirding_birds %>%
    filter(sovon_bird_reference %in%
      rings_to_add_megev$sovon_bird_reference) %>%
    select(
      sovon_bird_reference,
      sovon_bird_shorthand,
      metal_ring_missing,
      sovon_bird_date_begin,
      sovon_bird_date_end,
      everything()
    )
} else {
  message("'megev' is not present.")
}
```

We create now new rows based on action `klgev`. We set `sovon_bird_rings_changed` = 1, `metal_ring_missing` = FALSE and we leave `bird_shorthand` empty:

```{r show_rings_to_add_klgev}
rings_to_add_klgev <-
  obs_and_acts %>%
  filter(klgev == "klgev") %>%
  select(sovon_bird_reference, Datum) %>%
  left_join(crbirding_birds, by = "sovon_bird_reference") %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin < Datum) %>%
  filter(sovon_bird_date_begin == max(sovon_bird_date_begin)) %>%
  mutate(
    sovon_bird_date_begin = Datum,
    sovon_bird_date_end = case_when(
      sovon_bird_date_end == sovon_bird_date_begin ~ as.POSIXct.Date(NA),
      TRUE ~ sovon_bird_date_end
    )
  ) %>%
  select(-Datum) %>%
  mutate(
    sovon_bird_rings_changed = 1,
    metal_ring_missing = FALSE,
    sovon_bird_shorthand = NA_character_,
    sovon_bird_shorthand_pt = NA_character_
  ) %>%
  ungroup() %>%
  arrange(sovon_bird_reference)
rings_to_add_klgev
```

We add then these new rows, paying attention to change `bird_date_end` of previous rings:

```{r add_rings_klgev}
crbirding_birds <-
  crbirding_birds %>%
  bind_rows(rings_to_add_klgev) %>%
  arrange(
    sovon_bird_reference,
    sovon_bird_date_begin,
    sovon_bird_date_end
  ) %>%
  group_by(sovon_bird_reference, sovon_bird_date_end) %>%
  mutate(sovon_bird_date_end_updated = case_when(
    sovon_bird_date_begin == min(sovon_bird_date_begin) &
      sovon_bird_date_begin != max(sovon_bird_date_begin)
    ~ max(sovon_bird_date_begin),
    TRUE ~ sovon_bird_date_end
  )) %>%
  ungroup() %>%
  mutate(sovon_bird_date_end = sovon_bird_date_end_updated) %>%
  select(-sovon_bird_date_end_updated)
```

Show changes:

```{r show_changes_klgev}
crbirding_birds %>%
  filter(sovon_bird_reference %in%
    rings_to_add_klgev$sovon_bird_reference) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    sovon_bird_rings_changed,
    metal_ring_missing,
    everything()
  )
```

### Finalize `sovon_bird_rings_changed`

There are still rings where `sovon_bird_rings_changed` is empty:

```{r empty_sovon_bird_rings_changed}
crbirding_birds %>%
  filter(is.na(sovon_bird_rings_changed))
```

Show more details, in particular the actions of observations related to the birds the rings come from and at the same date as `sovon_bird_date_begin`:

```{r acts_crbirding_birds_sovon_bird_date_begin}
crbirding_birds %>%
  filter(is.na(sovon_bird_rings_changed)) %>%
  select(sovon_bird_reference, sovon_bird_date_begin) %>%
  left_join(obs_and_acts %>%
    select(Nummer, sovon_bird_reference, Datum, acts),
  by = c("sovon_bird_reference",
    "sovon_bird_date_begin" = "Datum"
  )
  ) %>%
  select_if(function(x) any(!is.na(x)))
```

As expected most of them are `vang` only actions linked to applying new color rings. 

The `veld` actions only are just field observations occurring the very same date of `vang` action and will be never used for mapping.

There are some `ziek` only actions:

```{r ziek_mapping_crbirding_birds}
crbirding_birds %>%
  filter(is.na(sovon_bird_rings_changed)) %>%
  select(sovon_bird_reference, sovon_bird_shorthand, sovon_bird_date_begin) %>%
  left_join(obs_and_acts %>%
    select(Nummer, sovon_bird_reference, Datum, acts, Opmerking),
  by = c("sovon_bird_reference",
    "sovon_bird_date_begin" = "Datum"
  )
  ) %>%
  filter(ziek == "ziek" & is.na(vangl)) %>%
  select_if(function(x) any(!is.na(x)))
```

These three rings have been added before as particular cases of ringing while being sick. Based on notes and previous observations, we can see that `FAAG` has been applied together with a new metal ring, i.e. `sovon_bird_rings_changed` = 2, while the other two don't mention any change of metal ring: `sovon_bird_rings_changed` = 1:

```{r map_sovon_bird_rings_changed_FAAG_YCAF_FHOV}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_rings_changed = case_when(
    sovon_bird_shorthand == "FAAG" ~ 2,
    sovon_bird_shorthand %in% c("YCAF", "FHOV") ~ 1,
    TRUE ~ sovon_bird_rings_changed
  ))
```

The combination of `vang` + `klweg` doesn't change anything in the mapping, as noticing the absence of color ring happens at the same time of applying a new one. We can assess whether the metal ring is still present by checking the value of `metal_ring_missing` for the very previous ring. We make now **two important assumptions**:

1. the ringer applies always a new metal ring if missing
2. the ringer doesn't change a metal ring if present

So, `vang` after a `meweg` or `vang` together with `meweg` are interpreted as applying a new metal ring, therefore `sovon_bird_rings_changed` = 2. Again, this is a consequence of the fact that the change of a metal ring is not coded by any action, but only described in informative notes.

Effects of this mapping:

```{r apply_sovon_bird_rings_changed_vang_final}
rings_without_sovon_bird_rings_changed <-
  crbirding_birds %>%
  filter(is.na(sovon_bird_rings_changed)) %>%
  select(sovon_bird_reference, sovon_bird_date_begin) %>%
  rename(date_ringing = sovon_bird_date_begin) %>%
  left_join(crbirding_birds, by = "sovon_bird_reference") %>%
  group_by(sovon_bird_reference) %>%
  filter(sovon_bird_date_begin < date_ringing) %>%
  filter(sovon_bird_date_begin == max(sovon_bird_date_begin)) %>%
  mutate(
    sovon_bird_rings_changed = case_when(
      is.na(sovon_bird_shorthand) ~ 1,
      !is.na(sovon_bird_shorthand) & isTRUE(metal_ring_missing) ~ 2,
      !is.na(sovon_bird_shorthand) & isFALSE(metal_ring_missing) ~ 1
    ),
    metal_ring_missing = FALSE
  ) %>%
  ungroup() %>%
  select(
    sovon_bird_reference,
    sovon_bird_rings_changed,
    date_ringing,
    metal_ring_missing
  ) %>%
  rename(sovon_bird_date_begin = date_ringing) %>%
  left_join(crbirding_birds %>%
    select(-c(
      sovon_bird_rings_changed,
      metal_ring_missing
    )),
  by = c(
    "sovon_bird_reference",
    "sovon_bird_date_begin"
  )
  ) %>%
  select(names(crbirding_birds)) %>%
  left_join(obs_and_acts %>%
    filter((vang == "vang" | vangl == "vangl") & meweg == "meweg") %>%
    select(sovon_bird_reference, Datum, meweg),
  by = c("sovon_bird_reference",
    "sovon_bird_date_begin" = "Datum"
  )
  ) %>%
  mutate(sovon_bird_rings_changed = case_when(
    sovon_bird_rings_changed == 1 & meweg == "meweg" ~ 2,
    TRUE ~ sovon_bird_rings_changed
  )) %>%
  select(names(crbirding_birds))
rings_without_sovon_bird_rings_changed
```

Update `crbirding_birds` based on the mapping above:

```{r update_crbirding_birds_vang}
crbirding_birds <-
  crbirding_birds %>%
  anti_join(rings_without_sovon_bird_rings_changed,
    by = c(
      "sovon_bird_reference",
      "sovon_bird_shorthand",
      "sovon_bird_shorthand_pt",
      "sovon_bird_date_begin",
      "sovon_bird_date_end"
    )
  ) %>%
  bind_rows(rings_without_sovon_bird_rings_changed) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

Rings still without `sovon_bird_rings_changed`:

```{r check_n_rings_without_sovon_bird_rings_changed}
crbirding_birds %>%
  filter(is.na(sovon_bird_rings_changed))
```

These are intermediate rings of birds which got three color rings:

```{r show_sequence_rings_triplets_or_more}
crbirding_birds %>%
  filter(sovon_bird_reference %in%
    (crbirding_birds %>%
      filter(is.na(sovon_bird_rings_changed)) %>%
      pull(sovon_bird_reference)))
```

All of them will get `sovon_bird_rings_changed` = 1:

```{r map_sovon_bird_rings_changed_intermediate_color_rings}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_rings_changed = case_when(
    is.na(sovon_bird_rings_changed) ~ 1,
    TRUE ~ sovon_bird_rings_changed
  ))
```

Check that all rings have a valid value (0, 1 or 2) for `sovon_bird_rings_changed`:

```{r check_mapping_complete_sovon_bird_rings_changed}
crbirding_birds %>%
  filter(is.na(sovon_bird_rings_changed)) %>%
  nrow() == 0 &
  all(unique(crbirding_birds$sovon_bird_rings_changed) %in%
    c(0, 1, 2))
```

We show a final overview. 

Birds getting color and metal ring at the same time and never ringed twice (only the first thousand are shown as this category includes more than 10 thousands rings):

```{r final_overview_crbiridng_birds_n_1}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(sovon_bird_rings_changed == 2) %>%
  filter(n == 1) %>%
  select(-n)
```

Birds ringed in two different moments (only the first thousand are shown as this category includes more than two thousands rings):

```{r final_overview_crbiridng_birds_n_2}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n == 2) %>%
  select(-n)
```

Birds ringed at three different moments:

```{r final_overview_crbiridng_birds_n_3}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n == 3) %>%
  select(-n)
```

Birds ringed at four different moments:

```{r final_overview_crbiridng_birds_n_4}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  filter(n == 4) %>%
  select(-n)
```

At Sept, 25 no birds are ringed more than four times:

```{r final_overview_crbiridng_birds_n_higher_than_4}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  filter(n >= 5) %>%
  select(-n)
```
### Map color of rings and inscriptions

The color of the ring should be added, if available, to `bird_shorthand`. It's important to highlight the fact that we have this information only for the most recent color ring of each bird. Following combinations are present, although the color of the inscription `InscriptieKleurCode` will be not mapped:

```{r color_combinations_inbo_db}
birds %>%
  group_by(RingKleurCode, InscriptieKleurCode) %>%
  count()
```

We first map the colors based on the official cr-birding values ( [cr-birding.org](http://www.cr-birding.org/node/112)):

```{r map_colors}
color_table <-
  color_table %>%
  mutate(crbirding_color = case_when(
    Code == "X" ~ "P",
    Code == "Z" ~ "N",
    TRUE ~ Code
  ))
color_table
```

We map the colors in `birds`:

```{r map_colors_birds}
birds <-
  birds %>%
  left_join(color_table %>%
    select(Code, crbirding_color),
  by = c("RingKleurCode" = "Code")
  ) %>%
  rename(ring_color = crbirding_color) %>%
  left_join(color_table %>%
    select(Code, crbirding_color),
  by = c("InscriptieKleurCode" = "Code")
  ) %>%
  rename(inscription_color = crbirding_color)
```

Effects of the mapping:

```{r result_mapping_colors_birds}
birds %>%
  distinct(ring_color, RingKleurCode, inscription_color, InscriptieKleurCode)
```

The `bird_shorthand` should be composed of:
1. color of the ring
3. inscription

For example, B-BBAP is a dark blue ring (see [issue #67](https://github.com/inbo/cr-birding/issues/67#issuecomment-643954772)).

We add ring color to `bird_shorthand`:

```{r add_colors_bird_shorthand}
crbirding_birds <-
  crbirding_birds %>%
  left_join(birds %>%
              select(sovon_bird_reference, ring_color),
            by = "sovon_bird_reference") %>%
  group_by(sovon_bird_reference) %>%
  filter(!is.na(sovon_bird_shorthand)) %>%
  mutate(ring_color = if_else(
    sovon_bird_date_begin == max(sovon_bird_date_begin),
    ring_color,
    NA_character_
  )) %>%
  bind_rows(crbirding_birds %>%
              filter(is.na(sovon_bird_shorthand))) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin) %>%
  mutate(
    sovon_bird_shorthand_color = case_when(
      !is.na(sovon_bird_shorthand) &
        !is.na(ring_color) ~ paste0(
          ring_color,
          "-",
          sovon_bird_shorthand
        ),
      TRUE ~ sovon_bird_shorthand
    ),
    sovon_bird_shorthand_pt_color = case_when(
      !is.na(sovon_bird_shorthand_pt) &
        !is.na(ring_color) ~ paste0(
          ring_color,
          "-",
          sovon_bird_shorthand_pt
        ),
      TRUE ~ sovon_bird_shorthand_pt
    )
  ) %>%
  ungroup()
```

As inscription color would be lost in cr-birding database, we add it as a note: `inscription color: ` followed by the EURING Exchange Code 2000+ standard color code:

```{r add_incsription_color_as_note}
crbirding_birds <- 
  crbirding_birds %>%
  # filter(str_sub(sovon_bird_shorthand_color, start = 2, end = 2) == "-") %>%
  left_join(birds %>%
              select(sovon_bird_reference, inscription_color),
            by = "sovon_bird_reference") %>%
  mutate(sovon_bird_notes = if_else(
    str_sub(sovon_bird_shorthand_color, start = 2, end = 2) == "-",
    if_else(is.na(sovon_bird_notes),
            str_c("inscription color: ", inscription_color, "."),
            str_c(sovon_bird_notes,
                  str_c("inscription color: ", inscription_color, "."),
                  sep = " ")),
    sovon_bird_notes))
```

Some examples from birds ringed thrice or more:

```{r show_example_colors}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 2) %>%
  select(
    sovon_bird_reference,
    ring_color,
    inscription_color,
    sovon_bird_shorthand_color,
    sovon_bird_shorthand_pt_color,
    sovon_bird_shorthand,
    sovon_bird_shorthand_pt,
    sovon_bird_notes,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    everything()
  )
```

Based on INBO experts knowledge, we can assign color of rings and inscriptions for a significant number of old, changed rings as well. All rings with a 4-letter long inscription where the third letter is an A are blue rings with white inscription (see https://github.com/inbo/cr-birding/issues/84#issuecomment-478938321).

Rings still not mapped satisfying this condition:

```{r check_n_color_old_rings_with_A_3rd_position}
old_rings_with_A <-
  crbirding_birds %>%
  filter(
    !str_detect(sovon_bird_shorthand_color, "\\-") &
      str_detect(sovon_bird_shorthand, "[A-Z]") &
      str_length(sovon_bird_shorthand) == 4 &
      str_sub(sovon_bird_shorthand, 3, 3) == "A"
  )
old_rings_with_A
```

Assign colors:

```{r assign_color_old_rings}
old_rings_with_A <-
  old_rings_with_A %>%
  mutate(
    sovon_bird_shorthand_color = paste0("B-",
                                        sovon_bird_shorthand_color),
    sovon_bird_shorthand_pt_color = paste0("B-",
                                           sovon_bird_shorthand_pt_color),
    sovon_bird_notes = if_else(is.na(sovon_bird_notes),
                               "inscription color: W.",
                               str_c(sovon_bird_notes,
                                     "inscription color: W.",
                                     sep = " "))
    ) %>%
  select(
    sovon_bird_reference,
    ring_color,
    sovon_bird_shorthand_color,
    sovon_bird_shorthand_pt_color,
    sovon_bird_shorthand,
    sovon_bird_shorthand_pt,
    sovon_bird_notes,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    everything()
  )
old_rings_with_A
```

Apply changes to `crbirding_birds`:

```{r merge_old_rings_with_A_to_crbirding_birds}
crbirding_birds <-
  crbirding_birds %>%
  anti_join(old_rings_with_A,
    by = c(
      "sovon_bird_shorthand",
      "sovon_bird_date_begin",
      "sovon_bird_date_end"
    )
  ) %>% 
  bind_rows(old_rings_with_A) %>%
  select(names(crbirding_birds)) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

Some examples from birds ringed thrice or more:

```{r show_example_colors_old_rings}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(n > 2) %>%
  select(
    sovon_bird_reference,
    ring_color,
    sovon_bird_shorthand_color,
    sovon_bird_shorthand_pt_color,
    sovon_bird_shorthand,
    sovon_bird_shorthand_pt,
    sovon_bird_notes,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    everything()
  )
```

Rings still without color:

```{r old_rings_without_color}
rings_color_still_unknown <-
  crbirding_birds %>%
  filter(!str_detect(sovon_bird_shorthand_color, "\\-"))
rings_color_still_unknown %>%
  select(
    sovon_bird_reference,
    ring_color,
    sovon_bird_shorthand_color,
    sovon_bird_shorthand_pt_color,
    sovon_bird_shorthand,
    sovon_bird_shorthand_pt,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    everything()
  )
```

Some ring inscriptions contain a `B` at third position:

```{r not_coloured_rings_B}
rings_color_still_unknown_with_B_third_pos <-
  rings_color_still_unknown %>%
  filter(str_sub(sovon_bird_shorthand, 3, 3) == "B")
rings_color_still_unknown_with_B_third_pos
```

INBO experts explained us that the rings containing a `B` at third position are *virtual* rings (see [issue #81#issuecomment-478954317](https://github.com/inbo/cr-birding/issues/81#issuecomment-478954317)). It has been suggested to remove them and not exporting them (see [issues #81#issuecomment-698964036](https://github.com/inbo/cr-birding/issues/81#issuecomment-698964036)). We remove them:

```{r remove_virtual_rings}
crbirding_birds <- 
  crbirding_birds %>%
  filter(!
    sovon_bird_reference %in% rings_color_still_unknown_with_B_third_pos$sovon_bird_reference
  )
```

And we remove related observations as well:

```{r remove_obs_of_virtual_rings}
obs_and_acts <-
  obs_and_acts %>%
  filter(!
    sovon_bird_reference %in% rings_color_still_unknown_with_B_third_pos$sovon_bird_reference)
```

The rings left are foreign rings.

```{r show_foreign_rings}
rings_color_still_unknown_foreign <-
  crbirding_birds %>%
  filter(!str_detect(sovon_bird_shorthand_pt_color, "\\-"))
rings_color_still_unknown_foreign
```

We assign fictive color to them to ease the submission to cr-birding database. We assign them blue as ring color (see [issue #84](https://github.com/inbo/cr-birding/issues/84#issuecomment-479443389)):

```{r solve_foreign_rings}
rings_color_still_unknown_foreign <-
  rings_color_still_unknown_foreign %>%
  mutate(
    sovon_bird_notes = ifelse(is.na(sovon_bird_notes),
      "foreign_color_ring.",
      str_c(sovon_bird_notes,
        "foreign_color_ring.",
        sep = " "
      )
    ),
    sovon_bird_shorthand_color = paste0(
      "B-",
      sovon_bird_shorthand_color
    ),
    sovon_bird_shorthand_pt_color = paste0(
      "B-",
      sovon_bird_shorthand_pt_color
    )
  )

crbirding_birds <-
  crbirding_birds %>%
  anti_join(rings_color_still_unknown_foreign,
    by = c(
      "sovon_bird_reference",
      "sovon_bird_shorthand",
      "sovon_bird_shorthand_pt",
      "sovon_bird_date_begin",
      "sovon_bird_date_end"
    )
  ) %>%
  bind_rows(rings_color_still_unknown_foreign) %>%
  arrange(
    sovon_bird_reference,
    sovon_bird_date_begin
  )
```

After mapping:

```{r show_changes_foreign_rings}
crbirding_birds %>%
  filter(sovon_bird_reference %in%
    rings_color_still_unknown_foreign$sovon_bird_reference) %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_shorthand_color,
    sovon_bird_shorthand_pt_color,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    sovon_bird_notes,
    everything()
  )
```

Did we assigned a color to all rings?

```{r check_presence_colors_in_bird_shorthand}
crbirding_birds %>%
  filter(is.na(sovon_bird_shorthand_color) |
           str_detect(sovon_bird_shorthand_pt_color, "\\-")) %>%
  nrow() == nrow(crbirding_birds)
```

We will overwrite the columns `sovon_bird_shorthand` and `sovon_bird_shorthand_pt` and remove the help columns  `sovon_bird_shorthand_color`, `sovon_bird_shorthand_pt_color` and `ring_color` at the end of the processing.

### Position of the color ring and inscription orientation

SOVON allows us to add two columns called `position_color_ring` and `direction_color_ring` in order to save:
1. the position of the color ring, left or right leg
2. the direction of the color ring's inscription, upwards or downwards 

In case of birds with one color ring only, SOVON doesn't save these fields. However, INBO experts find them important as 

> it can help a lot when identifying mistakes in reading. 

For more details, see [issue #61](https://github.com/inbo/cr-birding/issues/61#issuecomment-472862064).

As for colors, these informations are only available for the very last ring of each bird. 

Overview:

```{r overview_places}
birds %>%
  group_by(Plaats) %>%
  count()
```

We map all values of ring position and inscription reading direction:

```{r map_positions_and_directions}
ring_position_table <-
  ring_position_table %>%
  mutate(
    sovon_bird_ring_position = case_when(
      str_sub(Code, start = 1, end = 1) == "L" ~ "LB",
      str_sub(Code, start = 1, end = 1) == "R" ~ "RB",
      TRUE ~ NA_character_
    ),
    sovon_bird_ring_direction = case_when(
      str_sub(Code, start = -1, end = -1) == "D" ~ "D",
      str_sub(Code, start = -1, end = -1) == "U" ~ "U",
      TRUE ~ NA_character_
    )
  ) %>%
  select(
    Code,
    sovon_bird_ring_position,
    sovon_bird_ring_direction,
    everything()
  )
ring_position_table
```

where `LB` and  `RB` stay for _left tarsus_ and _right tarsus_ respectively.

We map the ring position and direction of ring inscription in `birds`:

```{r map_pos_direction_birds}
birds <-
  birds %>%
  left_join(ring_position_table %>%
    select(Code, sovon_bird_ring_position),
  by = c("Plaats" = "Code")
  ) %>%
  left_join(ring_position_table %>%
    select(Code, sovon_bird_ring_direction),
  by = c("Plaats" = "Code")
  )
```

Effects of the mapping:

```{r result_mapping_place}
birds %>%
  distinct(Plaats, sovon_bird_ring_position, sovon_bird_ring_direction)
```

Map these values from `birds` to `crbirding_birds`, taking into account that the fields `sovon_bird_ring_position` and `sovon_bird_ring_direction` are available only for the very last color ring of each bird:

```{r transfer_position_direction_to_crbirding_birds}
crbirding_birds <-
  crbirding_birds %>%
  filter(!is.na(sovon_bird_shorthand)) %>%
  left_join(birds %>%
    select(
      sovon_bird_reference,
      sovon_bird_ring_position,
      sovon_bird_ring_direction
    ),
  by = "sovon_bird_reference"
  ) %>%
  group_by(sovon_bird_reference) %>%
  mutate(
    sovon_bird_ring_position = if_else(
      sovon_bird_date_begin == max(sovon_bird_date_begin),
      sovon_bird_ring_position,
      NA_character_
    ),
    sovon_bird_ring_direction = if_else(
      sovon_bird_date_begin == max(sovon_bird_date_begin),
      sovon_bird_ring_direction,
      NA_character_
    )
  ) %>%
  ungroup() %>%
  bind_rows(crbirding_birds %>%
    filter(is.na(sovon_bird_shorthand))) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

Some examples from birds ringed more than twice:

```{r show_example_position}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  filter(n > 2) %>%
  select(
    sovon_bird_reference,
    sovon_bird_ring_position,
    sovon_bird_ring_direction,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    everything()
  ) %>%
  head(n = 100)
```

### Mapping of `sovon_bird_ring_number`

We can now use `metal_ring_missing` in order to add the metal ring number as `sovon_bird_ring_number`. Still, note that **only the very last metal ring number is saved in our color-oriented database**. We will use it in place of older metal rings. Experts could improve it in crbirding database later.

```{r add_sovon_bird_ring_number_to_crbirding_birds}
crbirding_birds <-
  crbirding_birds %>%
  left_join(birds %>%
    select(sovon_bird_reference, sovon_bird_ring_number) %>%
    rename(metal_ring_number = sovon_bird_ring_number),
  by = c("sovon_bird_reference")
  ) %>%
  group_by(sovon_bird_reference) %>%
  mutate(sovon_bird_ring_number = case_when(
    metal_ring_missing == FALSE ~ metal_ring_number,
    metal_ring_missing == TRUE ~ NA_character_
  )) %>%
  select(-metal_ring_number) %>%
  ungroup()
```

Show an example by selecting birds wich lost their own metal rings:

```{r example_apply_metalring}
crbirding_birds %>%
  filter(metal_ring_missing == TRUE) %>%
  select(sovon_bird_reference) %>%
  left_join(crbirding_birds, by = "sovon_bird_reference") %>%
  select(
    sovon_bird_reference,
    sovon_bird_date_begin,
    sovon_bird_date_end,
    sovon_bird_ring_number,
    metal_ring_missing,
    sovon_bird_rings_changed,
    everything()
  )
```

Add note about absence of metal ring based on original values in column `MetaaalringNummer` and the fields `sovon_bird_reference`, `sovon_bird_ring_number`:

```{r add_note_absence_metal_ring}
crbirding_birds <-
  crbirding_birds %>%
  left_join(birds %>%
    select(
      sovon_bird_reference,
      MetaalringNummer
    ),
  by = c("sovon_bird_reference")
  ) %>%
  mutate(sovon_bird_notes = case_when(
    is.na(sovon_bird_ring_number) &
      metal_ring_missing == FALSE &
      (MetaalringNummer %in% c("?", "onbekend") |
        is.na(MetaalringNummer)) ~
    ifelse(is.na(sovon_bird_notes),
      "bird_ring_number not available.",
      str_c(sovon_bird_notes, "bird_ring_number not available.", sep = " ")
    ),
    is.na(sovon_bird_ring_number) &
      metal_ring_missing == FALSE &
      MetaalringNummer == "none" ~
    ifelse(is.na(sovon_bird_notes),
      "bird_ring_number not present.",
      str_c(sovon_bird_notes, "bird_ring_number not present.", sep = " ")
    ),
    TRUE ~ sovon_bird_notes
  )) %>%
  select(-MetaalringNummer)
```

Notice that in these cases the ring number is unknown although we are sure it exists. We left `sovon_bird_ring_number` empty as suggested by SOVON experts (see [comment on issue #100](https://github.com/inbo/cr-birding/issues/100#issuecomment-481217404)) without modifying the value of `bird_rings_changed`.

This is how `sovon_bird_notes` has been updated, limited to birds where changes could happen:

```{r updated_bird_notes}
crbirding_birds %>%
  filter(is.na(sovon_bird_ring_number)) %>%
  filter(metal_ring_missing == FALSE) %>%
  select(sovon_bird_reference) %>%
  left_join(crbirding_birds,
    by = "sovon_bird_reference"
  ) %>%
  select(-sovon_bird_shorthand) %>%
  arrange(sovon_bird_reference, sovon_bird_date_begin)
```

### Euring

Present values:

```{r euring_in_db}
birds %>% distinct(EuringCode)
```

We assign the euring codes by applying the following mapping:

```{r sovon_bird_euring}
crbirding_birds <-
  crbirding_birds %>%
  left_join(birds %>%
    distinct(sovon_bird_reference, EuringCode),
  by = "sovon_bird_reference"
  ) %>%
  mutate(sovon_bird_euring = recode(EuringCode,
    "5920" = "05920",
    "5910" = "05910",
    "5926" = "05926",
    "5922" = "06009", # 5922 used for hybrid gulls
    "05922" = "06009", # 05922 used for hybrid gulls
    "zz" = "04560", # pied avocets
    "zzz" = "00720" # great cormorants
  )) %>%
  ungroup()
```

Effects of mapping:

```{r comparison_euring}
crbirding_birds %>%
  distinct(EuringCode, sovon_bird_euring)
```

We remove the help column `EuringCode` from `crbirding_birds`:

```{r remove_euringCode}
crbirding_birds <-
  crbirding_birds %>%
  select(-EuringCode)
```

### Scheme

Actual values:

```{r scheme}
birds %>% distinct(MetaalringLandCode)
```

We apply the following mapping:

```{r scheme_mapping}
crbirding_birds <-
  crbirding_birds %>%
  left_join(birds %>%
    select(sovon_bird_reference, MetaalringLandCode),
  by = "sovon_bird_reference"
  ) %>%
  mutate(sovon_bird_scheme = recode(MetaalringLandCode,
    "BE" = "BLB",
    "FR" = "FRP",
    "NL" = "NLA",
    "PT" = "POL",
    "UK" = "GBT"
  )) %>%
  mutate(sovon_bird_scheme = case_when(
    is.na(sovon_bird_ring_number) ~ NA_character_,
    TRUE ~ sovon_bird_scheme
  ))
```

Effects of mapping:

```{r comparison_scheme}
crbirding_birds %>%
  filter(!is.na(sovon_bird_ring_number)) %>%
  distinct(MetaalringLandCode, sovon_bird_scheme)
```

We remove the help column `MetaalringLandCode` from `crbirding_birds`:

```{r remove_MetaalringLandCode}
crbirding_birds <-
  crbirding_birds %>%
  select(-MetaalringLandCode)
```

### Bird sex

Bird sex is translated to English. Letter `M` (Dutch word _mannetje_) will not change so no need to convert it:

```{r sovon_bird_sex}
crbirding_birds <-
  crbirding_birds %>%
  left_join(birds %>%
    select(sovon_bird_reference, GeslachtCode),
  by = "sovon_bird_reference"
  ) %>%
  mutate(sovon_bird_sex = recode(GeslachtCode,
    "V" = "F", ## V(rouwtje) ->F(emale)
    "O" = "U" ## O(nbekend) ->  U(nknown)
  ))
```

Effects of mapping:

```{r comparison_sex}
crbirding_birds %>% distinct(GeslachtCode, sovon_bird_sex)
```

We remove the help column `GeslachtCode` from `crbirding_birds`:

```{r remove_GeslachtCode}
crbirding_birds <-
  crbirding_birds %>%
  select(-GeslachtCode)
```

### Bird age ringing

For mapping the age while applying color rings, we have to follow the Euring standard: see [online pdf document](https://euring.org/files/documents/E2000PLUSExchangeCodev1161.pdf) at page 14. 

Bird age at the moment of any observation can be found in column `LeeftijdCode` of `obs_and_acts`. Values present:

```{r show_present_LeeftijdCode}
obs_and_acts %>%
  distinct(LeeftijdCode)
```

We apply a recoding in order to standardize INBO's vocabulary to the EURING standard:

```{r map_age_code}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_bird_age_obs = recode(LeeftijdCode,
    "PU" = "1",
    "AD" = "A",
    "J1" = "3",
    "I4" = "5",
    "I3" = "7",
    "I2" = "9",
    "I5" = "B",
    .missing = NA_character_
  ))
```

The age can be added to `crbirding_birds` by matching `sovon_bird_date_begin` of `crbirding_birds` and `Datum` of `obs_and_acts` for each `sovon_bird_reference`. First we check that this strategy ends up with a one-to-one relation:

```{r check_uniqueness_approach}
crbirding_birds %>%
  left_join(obs_and_acts %>%
    ## remove rows with Datum equal to NA or field observations
    filter(!is.na(sovon_bird_age_obs) &
      (!is.na(vang) | !is.na(vangl) |
        !is.na(rngkl) | !is.na(rngme) |
        !is.na(klgev) | !is.na(klweg) |
        !is.na(meweg) |
        !is.na(ziek) # for FHOV,YCAF,FAAG
      )),
  by = c("sovon_bird_reference",
    "sovon_bird_date_begin" = "Datum"
  )
  ) %>%
  nrow() == nrow(crbirding_birds)
```

Apply mapping:

```{r apply_mapping_bird_age_ringing}
crbirding_birds <-
  crbirding_birds %>%
  left_join(obs_and_acts %>%
    ## remove rows with Datum equal to NA or field observations
    filter(!is.na(sovon_bird_age_obs) & 
      (!is.na(vang) | !is.na(vangl) |
        !is.na(rngkl) | !is.na(rngme) |
        !is.na(klgev) | !is.na(klweg) |
        !is.na(meweg) |
        !is.na(ziek) # for FHOV,YCAF,FAAG
      )) %>%
    select(sovon_bird_reference, Datum, sovon_bird_age_obs),
  by = c("sovon_bird_reference",
    "sovon_bird_date_begin" = "Datum"
  )
  )
```

It can happen that there are birds without age while ringing. This happens typically while changing a ring, i.e. birds with more than one ring:

```{r n_rings_of_birds_without_ring}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(sovon_bird_reference %in% 
           (crbirding_birds %>%
           filter(is.na(sovon_bird_age_obs)) %>%
           pull(sovon_bird_reference))) %>%
  rename(n_rings = n) %>%
  group_by(n_rings) %>%
  count()
```

This is considered not a problem by SOVON ITer (see [113#issuecomment-605295649](https://github.com/inbo/cr-birding/issues/113#issuecomment-605295649)).

Birds without `sovon_bird_age_obs` while ringing once should be checked:

```{r birds_ringed_once_without_sovon_bird_age_obs}
crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  add_tally() %>%
  ungroup() %>%
  filter(sovon_bird_reference %in% 
           (crbirding_birds %>%
              filter(is.na(sovon_bird_age_obs)) %>%
              pull(sovon_bird_reference))) %>%
  rename(n_rings = n) %>%
  filter(n_rings == 1)
```

We rename `sovon_bird_age_obs` to `sovon_bird_age_ringing` in `crbirding_birds`:

```{r rename_sovon_bird_age_obs}
crbirding_birds <-
  crbirding_birds %>%
  rename(sovon_bird_age_ringing = sovon_bird_age_obs)
```

As examples, the bird age mapping of birds (color) ringed thrice or more:

```{r show_example_bird_age_birds_ringed_thrice}
bird_reference_triplet <-
  crbirding_birds %>%
  group_by(sovon_bird_reference) %>%
  count() %>%
  filter(n >= 3) %>%
  pull(sovon_bird_reference)

crbirding_birds %>%
  select(
    sovon_bird_reference,
    sovon_bird_shorthand,
    sovon_bird_date_begin,
    sovon_bird_age_ringing
  ) %>%
  filter(sovon_bird_reference %in% bird_reference_triplet)
```

### Bird ID

Bird identifiers will be provided by SOVON. `NA` is given:

```{r birdID}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_id = NA)
```

### Bird BTO

Bird BTO will be provided by SOVON.  `NA` is given:

```{r bird_bto}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_bto = NA)
```

### Bird name

Some birds, the ones with a GPS tracker, have a name saved in [this file](https://github.com/inbo/bird-tracking/blob/8a7a2c91d54b486d90a97f85cfaa30cede67315d/data/interim/individuals.csv). The names are saved in column `individual_remarks`:

```{r read_names}
uvabits_file <- "https://raw.githubusercontent.com/inbo/bird-tracking/8a7a2c91d54b486d90a97f85cfaa30cede67315d/data/interim/individuals.csv"
uvabits_names <-
  read_csv(uvabits_file,
    col_types = cols(
      .default = col_character(),
      individual_id = col_double(),
      mass = col_double(),
      track_session_id = col_double(),
      device_info_serial = col_double(),
      tracker_id = col_double(),
      track_session_start_date = col_datetime(format = ""),
      track_session_end_date = col_datetime(format = ""),
      track_session_start_latitude = col_double(),
      track_session_start_longitude = col_double()
    )
  )
uvabits_names <- 
  uvabits_names %>%
  select(individual_id, ring_number, individual_remarks, everything())
uvabits_names
```

We couple the birds to the correspondent names by color ring:

```{r coupling_bird_reference_name}
bird_ref_uvabits <-
  crbirding_birds %>%
  left_join(uvabits_names %>%
    select(colour_ring, individual_remarks) %>%
    filter(!is.na(colour_ring)),
  by = c("sovon_bird_shorthand" = "colour_ring")
  ) %>%
  rename(sovon_bird_name = individual_remarks) %>%
  filter(!is.na(sovon_bird_name)) %>%
  select(sovon_bird_reference, sovon_bird_name)
```

And we assign the names by joining on `sovon_bird_reference`:

```{r add_names}
crbirding_birds <-
  crbirding_birds %>%
  left_join(bird_ref_uvabits, by = "sovon_bird_reference")
```

The complete mapping of names:

```{r example_bird_names}
crbirding_birds %>%
  filter(!is.na(sovon_bird_name)) %>%
  select(sovon_bird_reference, sovon_bird_name, sovon_bird_shorthand)
```

### Bird birth year

This field will be filled by SOVON. `NA` is given:

```{r bird_birth_year}
crbirding_birds <-
  crbirding_birds %>%
  mutate(sovon_bird_birth_year = NA)
```

## Finalize user data

We have now sufficient information to define the role of any user in `crbirding_users`.

### Add user role to `crbirding_users`

Based on `sovon_bird_date_begin` we can retrieve the observations linked to applying rings. We retrieve the ringers by joining `crbirding_users` and `obs_and_acts` by `WaarnemerNummer`:

```{r find_ringers}
ringers_number <-
  crbirding_birds %>%
  select(sovon_bird_reference, sovon_bird_date_begin) %>%
  left_join(obs_and_acts %>%
    ## remove rows with Datum equal to NA or field observations
    filter(!is.na(sovon_bird_age_obs) &
      is.na(veld) & is.na(dood) & is.na(br) &
      (!is.na(vang) | !is.na(vangl) |
        !is.na(rngkl) | !is.na(rngme) |
        !is.na(klgev) | !is.na(klweg) |
        !is.na(meweg) |
        !is.na(ziek) # for FHOV,YCAF,FAAG
      )) %>%
    select(sovon_bird_reference, Datum, WaarnemerNummer),
  by = c("sovon_bird_reference", "sovon_bird_date_begin" = "Datum")
  ) %>%
  filter(!is.na(WaarnemerNummer)) %>%
  distinct(WaarnemerNummer) %>%
  pull(WaarnemerNummer)
message(paste(
  "Number of ringers:",
  length(ringers_number)
))
```

We assign a `R` (ringer) to them, `O` (observer) otherwise:

```{r assign_ringer}
crbirding_users <-
  crbirding_users %>%
  mutate(user_role = ifelse(user_reference %in% ringers_number,
    "R", "O"
  ))
```

Number of ringers and observers:

```{r summary_n_ringers_observers}
crbirding_users %>%
  group_by(user_role) %>%
  count()
```

## Save modified temporary observation data

We overwrite the temporary observation data based on the added columns:

```{r overwrite_temp_obs_data}
write_tsv(
  obs_and_acts,
  path = here::here("data", "interim", "obs_and_actions.tsv"),
  na = "",
  append = FALSE
)
```

## Save modified ring position and direction table

We save the updated table containing color ring position and inscription direction in `data/interim`:

```{r save_updated_pos_dir_table}
write_tsv(
  ring_position_table,
  path = here::here("data", "interim", "ring_position_table.tsv"),
  na = "",
  append = FALSE
)
```

## Save final user data

The desired order of columns in `crbirding_users`:

```{r cols_order_users_final}
cr_users_cols <- c(
  "user_id", "user_reference", "user_email", "user_first_name",
  "user_last_name", "user_address", "user_postal_code", "user_place",
  "user_country", "user_language", "user_role"
)
```

Are all required columns present?

```{r check_presence_required_cols_users}
all(cr_users_cols %in% names(crbirding_users)) &
  length(cr_users_cols) == ncol(crbirding_users)
```

We overwrite `crbirding_users.csv` with added information:

```{r overwrite_crbirding_users}
write_csv(crbirding_users,
  path = here::here("data", "processed", "crbirding_users.csv"),
  na = ""
)
```

## Save final ring data

We overwrite the columns `sovon_bird_shorthand` and `sovon_bird_shorthand_pt` and remove the help columns `metal_ring_missing`, `sovon_bird_shorthand_color`, `sovon_bird_shorthand_pt_color` and `color_combination`:

```{r remove_cols_with_color_after_mapping}
crbirding_birds <-
  crbirding_birds %>%
  mutate(
    sovon_bird_shorthand = sovon_bird_shorthand_color,
    sovon_bird_shorthand_pt = sovon_bird_shorthand_pt_color
  ) %>%
  select(-c(
    ends_with("color"),
    starts_with("color"),
    metal_ring_missing
  ))
```

SOVON is interested in color ring versions with a dot if dots are present, even if their purpose is just improving readability. We can overwrite `sovon_bird_shorthand` with control column `sovon_bird_shorthand_pt`:

```{r remove_col_bird_shorthand_pt}
crbirding_birds <-
  crbirding_birds %>%
  select(-sovon_bird_shorthand) %>%
  rename(sovon_bird_shorthand = sovon_bird_shorthand_pt)
```

Remove prefix `sovon_`:

```{r remove prefix_sovon_birds}
names(crbirding_birds) <- str_remove_all(names(crbirding_birds),
  pattern = "sovon_"
)
names(crbirding_birds)
```

Remove time and timezone information from `bird_date_begin` and `bird_date_end`:

```{r remove_time_and_tz}
crbirding_birds <-
  crbirding_birds %>%
  mutate(bird_date_begin = as.Date(bird_date_begin),
         bird_date_end = as.Date(bird_date_end))
```

The desired order of columns in `crbirding_birds`:

```{r cols_order_birds}
cr_birds_cols <- c(
  "bird_id", "bird_reference", "bird_euring", "bird_bto",
  "bird_shorthand", "bird_scheme", "bird_ring_number", "bird_name",
  "bird_sex", "bird_birth_year", "bird_date_begin", "bird_date_end",
  "bird_rings_changed", "bird_age_ringing", "bird_ring_position",
  "bird_ring_direction", "bird_notes"
)
```

Are all required columns present?

```{r check_presence_required_cols_birds}
all(cr_birds_cols %in% names(crbirding_birds)) &
  length(cr_birds_cols) == ncol(crbirding_birds)
```

Set column order:

```{r get_right_order_cols_birds}
crbirding_birds <-
  crbirding_birds %>%
  select(all_of(cr_birds_cols))
```

Preview data:

```{r final_preview_birds}
crbirding_birds %>% head(n = 100)
```

Save finale birds info as text file (comma separated value):

```{r write_bird_data_txt}
write_csv(
  crbirding_birds,
  path = here::here("data", "processed", "crbirding_birds.csv"),
  na = ""
)
```
