==> src/4_observations.Rmd <==
# Map observation data

## Read data

### Read temporary observation data

We start from the temporary observation data saved as TSV file in `data\interim`:

```{r import_temp_observations}
obs_and_acts <- read_tsv(
  here::here("data", "interim", "obs_and_actions.tsv"),
  col_types = cols(
    .default = col_character(),
    sovon_bird_reference = col_double(),
    Nummer = col_double(),
    Datum = col_datetime(format = ""),
    WaarnemerNummer = col_double(),
    PlaatsLengtegraadGraden = col_double(),
    PlaatsLengtegraadMinuten = col_double(),
    PlaatsLengtegraadSeconden = col_double(),
    PlaatsBreedtegraadGraden = col_double(),
    PlaatsBreedtegraadMinuten = col_double(),
    PlaatsBreedtegraadSeconden = col_double(),
    BevestigingDatum = col_datetime(format = ""),
    AanmaakDatum = col_datetime(format = ""),
    WijzigDatum = col_datetime(format = "")
  )
)
```

### Read action data

Import action codes and meaning:

```{r import_actions_meaning2}
actions_meaning <- read_tsv(here::here("data", "input", "actions_meaning.tsv"))
```

### Read processed ring data

We import finalized ring data from `./data/processed`:

```{r import_final_ring_data}
crbirding_birds <- read_csv(
  here::here("data", "processed", "crbirding_birds.csv"),
  col_types = cols(
    .default = col_character(),
    bird_id = col_logical(),
    bird_reference = col_double(),
    bird_bto = col_logical(),
    bird_birth_year = col_logical(),
    bird_date_begin = col_datetime(format = ""),
    bird_date_end = col_datetime(format = "")
  )
)
```

### Read temporary ring data

We will need some columns from original INBO ring data. We import the temporary ring data as well:

```{r import_birds_from_INBO_temp_data}
birds <- read_tsv(
  here::here("data", "interim", "birds.tsv"),
  col_types = cols(
    .default = col_character()
  )
)
```

### Read ring position data

We import mapping of ring position and inscription reading direction:

```{r read_updated_ring_position_table}
ring_position_table <- read_tsv(
  here::here("data", "interim", "ring_position_table.tsv"),
  na = "",
  col_types = cols(
    .default = col_character(),
    Aktief = col_logical()
  )
)
```

### Read processed user data

We import finalized user data from `./data/processed`:

```{r import_final_user_data}
crbirding_users <- read_csv(
  here::here("data", "processed", "crbirding_users.csv"),
  col_types = cols(
    .default = col_character(),
    user_id = col_logical(),
    user_reference = col_double()
  )
)
```

### Read list workers in bird shelters

```{r import_workers}
workers_bird_shelters <- read_tsv(
  here::here("data", "input", "workers_in_bird_shelters.tsv"),
  col_types = cols(
    .default = col_character(),
    user_id = col_logical(),
    user_reference = col_double()
  )
)
```

## Map color observation data

### Extract action codes

Actions present in `obs_and_acts` :

```{r acts_in_obs_and_acts}
acts <- actions_meaning$Code
acts <- acts[acts %in% names(obs_and_acts)]
acts
```

Combinations of actions present in observations:

```{r act_combinations}
action_combinations <-
  obs_and_acts %>%
  select(acts) %>%
  distinct()
action_combinations
```

### Observation ID

The field `sovon_observation_id` is left to SOVON:

```{r add_obs_id}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_id = NA)
```

### Observation reference

The observation reference is an unique identifier assigned to each observation. This field exists already: `Nummer`.

```{r add_obs_ref}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_reference = Nummer)
```

### Observation date

The date is saved in column `Datum`. We copy it in required column `sovon_observation_date`:

```{r obs_date}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_date = Datum)
```

### Observation time

There is no observation time in `obs_and_acts`. `NA` is given to `sovon_observation_time`:

```{r obs_time}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_time = NA)
```

### Observation latitude

Observations with missing or partially missing latitude information:

```{r nas_breedte_fields}
obs_missing_latitude_longitude <-
  obs_and_acts %>%
  filter(is.na(PlaatsBreedtegraadGraden) |
    is.na(PlaatsBreedtegraadMinuten) |
    is.na(PlaatsBreedtegraadSeconden) |
    is.na(PlaatsLengtegraadGraden) |
    is.na(PlaatsLengtegraadMinuten) |
    is.na(PlaatsLengtegraadSeconden)) %>%
  select(
    Nummer,
    KleurringNummer,
    Datum,
    starts_with("PlaatsBreedtegraad"),
    starts_with("PlaatsLengtegraad"),
    PlaatsGemeente,
    PlaatsToponym
  )
obs_missing_latitude_longitude
```

The absolute observation latitude should be converted to decimal degrees:

```{r obs_latitude}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_lat = case_when(
      !is.na(PlaatsBreedtegraadGraden) &
        !is.na(PlaatsBreedtegraadMinuten) &
        !is.na(PlaatsBreedtegraadSeconden) ~ conv_unit(
        str_c(PlaatsBreedtegraadGraden,
          PlaatsBreedtegraadMinuten,
          PlaatsBreedtegraadSeconden,
          sep = " "
        ),
        from = "deg_min_sec",
        to = "dec_deg"
      ),
      !is.na(PlaatsBreedtegraadGraden) &
        !is.na(PlaatsBreedtegraadMinuten) &
        is.na(PlaatsBreedtegraadSeconden) ~ conv_unit(
        str_c(PlaatsBreedtegraadGraden,
          PlaatsBreedtegraadMinuten,
          0,
          sep = " "
        ),
        from = "deg_min_sec",
        to = "dec_deg"
      ),
      !is.na(PlaatsBreedtegraadGraden) &
        is.na(PlaatsBreedtegraadMinuten) &
        is.na(PlaatsBreedtegraadSeconden) ~ conv_unit(
        str_c(PlaatsBreedtegraadGraden, 0, 0, sep = " "),
        from = "deg_min_sec",
        to = "dec_deg"
      ),
      TRUE ~ NA_character_
    )
  )
```

By field `PlaatsBreedtegraadRichtingCode` we can know whether the observation took place in the northern or southern hemisphere. In case of southern hemisphere, the latitude value should be negative.

Values of `PlaatsBreedtegraadRichtingCode` and number of observations:

```{r values_of_PlaatsBreedtegraadRichtingCode}
obs_and_acts %>%
  group_by(PlaatsBreedtegraadRichtingCode) %>%
  count()
```

If present, the observations with valid `PlaatsBreedtegraadGraden` but without `PlaatsBreedtegraadRichtingCode` should be further evaluated:

```{r obs_without_N_or_Z}
obs_and_acts %>%
  filter(is.na(PlaatsBreedtegraadRichtingCode) &
    !is.na(PlaatsBreedtegraadGraden)) %>%
  select(Nummer, starts_with("PlaatsBreedte"), PlaatsGemeente, PlaatsToponym)
```

Assign negative latitude if `PlaatsBreedtegraadRichtingCode` is equal to `S`:

```{r map_N_S}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_lat = case_when(
      PlaatsBreedtegraadRichtingCode == "N" ~ sovon_observation_lat,
      PlaatsBreedtegraadRichtingCode == "S" & !is.na(sovon_observation_lat) ~
      paste0("-", sovon_observation_lat),
      TRUE ~ sovon_observation_lat
    )
  )
```

### Observation longitude

Absolute value of the observation longitude should be converted to decimal degrees:

```{r obs_longitude}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_lng = case_when(
      !is.na(PlaatsLengtegraadGraden) &
        !is.na(PlaatsLengtegraadMinuten) &
        !is.na(PlaatsLengtegraadSeconden) ~ conv_unit(
        str_c(PlaatsLengtegraadGraden,
          PlaatsLengtegraadMinuten,
          PlaatsLengtegraadSeconden,
          sep = " "
        ),
        from = "deg_min_sec",
        to = "dec_deg"
      ),
      !is.na(PlaatsLengtegraadGraden) &
        !is.na(PlaatsLengtegraadMinuten) &
        is.na(PlaatsLengtegraadSeconden) ~ conv_unit(
        str_c(PlaatsLengtegraadGraden,
          PlaatsLengtegraadMinuten,
          0,
          sep = " "
        ),
        from = "deg_min_sec",
        to = "dec_deg"
      ),
      !is.na(PlaatsLengtegraadGraden) &
        is.na(PlaatsLengtegraadMinuten) &
        is.na(PlaatsLengtegraadSeconden) ~ conv_unit(
        str_c(PlaatsLengtegraadGraden, 0, 0, sep = " "),
        from = "deg_min_sec",
        to = "dec_deg"
      ),
      TRUE ~ NA_character_
    )
  )
```

By field `PlaatsLengtegraadRichtingCode` we can know whether the observation took place in the western or eastern hemisphere. In case of western hemisphere, the longitude value should be negative.

Values of `PlaatsLengtegraadRichtingCode` and number of observations:

```{r values_of_PlaatsLengtegraadRichtingCode}
obs_and_acts %>%
  group_by(PlaatsLengtegraadRichtingCode) %>%
  count()
```

If present, the observations with valid `PlaatsLengtegraadGraden` but without `PlaatsLengtegraadRichtingCode` should be further evaluated:

```{r obs_without_E_or_W}
obs_and_acts %>%
  filter(is.na(PlaatsLengtegraadRichtingCode) &
    !is.na(PlaatsLengtegraadGraden)) %>%
  select(Nummer, starts_with("PlaatsLengte"), PlaatsGemeente, PlaatsToponym)
```

Assign negative longitude if `PlaatsLengtegraadRichtingCode` is equal to `W`:

```{r map_E_W}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_lng = case_when(
      PlaatsLengtegraadRichtingCode == "E" ~ sovon_observation_lng,
      PlaatsLengtegraadRichtingCode == "W" & !is.na(sovon_observation_lng) ~ paste0("-", sovon_observation_lng),
      TRUE ~ sovon_observation_lng
    )
  )
```

### Observation location

Aggregate information about observation location. We follow the following structure: `PlaatsGemeente` [+ `, ` + `PlaatsToponym` [+ `: ` + `PlaatsToponymDetail`]]:

```{r obs_location}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_location = ifelse(!is.na(PlaatsToponym),
      str_c(PlaatsGemeente,
        PlaatsToponym,
        sep = ", "
      ),
      paste(PlaatsGemeente)
    )
  ) %>%
  mutate(
    sovon_observation_location = ifelse(!is.na(PlaatsToponymDetail),
      str_c(sovon_observation_location,
        PlaatsToponymDetail,
        sep = ": "
      ),
      paste(sovon_observation_location)
    )
  )
```

### Check bird

The field `sovon_check_bird` is provided by SOVON. `NA` is given:

```{r check_bird}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_check_bird = NA)
```

### User ID

The field `sovon_user_id` is left to SOVON:

```{r add_user_id}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_user_id = NA)
```

### User reference

The field `sovon_user_reference` links observations to users and it is equal to field `user_reference` in user data:

```{r add_user_reference}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_user_reference = WaarnemerNummer)
```

### Observer

The field `sovon_melder` is equal to `user_first_name` and `user_last_name` in `users`:

```{r obs_sovon_melder}
obs_and_acts <-
  obs_and_acts %>%
  left_join(
    crbirding_users %>%
      select(user_reference, user_first_name, user_last_name),
    by = c("WaarnemerNummer" = "user_reference")
  ) %>%
  mutate(sovon_melder = case_when(
    is.na(user_first_name) & !is.na(user_last_name) ~ user_last_name,
    !is.na(user_first_name) & is.na(user_last_name) ~ user_first_name,
    !is.na(user_first_name) & !is.na(user_last_name) ~ str_c(user_first_name, user_last_name, sep = " "),
    is.na(user_first_name) & is.na(user_last_name) ~ NA_character_
  )) %>%
  select(-c(user_first_name, user_last_name))
```

### Observer e-mail

The field `sovon_melder_email` is equal to `user_email` in `users`:

```{r obs_sovon_melder_email}
obs_and_acts <-
  obs_and_acts %>%
  left_join(
    crbirding_users %>%
      select(user_reference, user_email),
    by = c("WaarnemerNummer" = "user_reference")
  ) %>%
  mutate(sovon_melder_email = user_email)
```

### Reporter - ringer number

The field `sovon_melder_ringersnummer` is left to SOVON:

```{r sovon_melder_ringersnummer}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_melder_ringersnummer = NA)
```

### Ring number

The field `sovon_ring_number` contains information about the metal ring number, in the same format as in field `bird_ring_number` of `crbirding_birds`. As explained in the mapping of field `bird_ring_number` of `crbirding_birds`, INBO database is a color ring database. This field is not accurately mapped and we will leave it empty.

```{r add_ring_number}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_ring_number = NA_character_)
```

### Observation status

Adding the status (based on EURING) information arises by the need of mapping the observations with action code `br` and `vang`:

```{r breed_show}
actions_meaning %>%
  filter(Code %in% c("br", "vang"))
```

The actions `br` and `vang` refer to status `breeding` in EURING system. Based on the scheme at page 16 of [EURING Exchange Code 2000+](https://euring.org/files/documents/E2000PLUSExchangeCodev1161.pdf) document and based on discussion in [this issue](https://github.com/inbo/cr-birding/issues/114), we define the following mapping for field `status_full_grown_bird`:

```{r obs_status}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_status_full_grown_bird = case_when(
    !is.na(br) | !is.na(vang) ~ "N",
    LeeftijdCode == "PU" ~ "-",
    TRUE ~ "U"
  ))
```

### Observation condition

The SOVON field `observation_condition` is NOT mapped following the [EURING Exchange Code 2000+](https://euring.org/files/documents/E2000PLUSExchangeCodev1161.pdf) document. SOVON experts provided us the following list of conditions (see [#126](https://github.com/inbo/cr-birding/issues/126)):

- `1`: healthy
- `2`: sick/wounded
- `3`: dead (less than week)
- `4`: dead (more than week)
- `5`: dead (unknown if freshly dead)
- `6`: released after rehabilitation

So, we do an initial mapping of following actions as follows (in order of priority):

1. `dood`, `klgev`, `megev`: `observation_condition` = `5` 
2. `ziek` (or observations of bird shelter workers): `observation_condition` = `2`
3. `rngkl`, `rngme`, `vang`, `vangl`, `veld`, `me`, `meweg`, `klweg`, `br`: `observation_condition` = `1`

```{r obs_condition}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_condition = case_when(
    dood == "dood" ~ 5,
    klgev == "klgev" ~ 5,
    ziek == "ziek" | 
      sovon_user_reference %in% workers_bird_shelters$user_reference ~ 2,
    !is.na(rngkl) | !is.na(rngme) | !is.na(vang) | !is.na(vangl) |
    !is.na(veld) | !is.na(me) | !is.na(meweg) | !is.na(klweg) | !is.na(br) ~ 1
  ))
```

Releasing after rehabilitation can be detected by the notes in field `Opmerking` for the observations of bird shelter workers. This is due to the fact that some of these observations are linked to ringing actions, other are linked to `veld` actions thus making the detection of release by actions troublesome.

Assign condition value `6` to the release after rehabilitation:

```{r assign_6_release_after_rehabilitation}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_condition = if_else(
    # observation of a bird shelter worker
    sovon_user_reference %in% workers_bird_shelters$user_reference & 
      # bird is not dead
      is.na(dood) & 
      # no empty note
      !is.na(Opmerking) & 
      # note mentions the release in Dutch
      str_detect(Opmerking, pattern = "gelost|vrijlating") & 
      # note does not mention the unrelease
      !str_detect(Opmerking, pattern = "niet gelost"),
    6, sovon_observation_condition
  ))
```

Overview of notes and actions of observations with `sovon_observation_condition` = 6 (released after rehabilitation):

```{r show_obs_condition_6}
obs_and_acts %>%
  filter(sovon_observation_condition == 6) %>%
  select(
    Opmerking,
    acts
  ) %>%
  select_if(function(x) any(!is.na(x))) %>%
  distinct()
```

Preview of mapping of `sovon_observation_condition`: 

```{r preview}
obs_and_acts %>%
  select(
    sovon_observation_reference,
    sovon_bird_reference,
    sovon_observation_date,
    sovon_observation_condition,
    acts
  ) %>%
  head(n = 100)
```

### MRI: metal ring information

Metal ring information is an integer between 0 and 9. A table can be found at page 8 of [EURING Exchange Code 2000+](https://euring.org/files/documents/E2000PLUSExchangeCodev1161.pdf) document.

However, SOVON uses this field to map the observations referring to the applying of any kind of ring, either color or metal, as their database is purely focussed on color rings and due to a missing field in the EURING Exchange Code 2000+ specific for color rings.

INBO experts explained us that all color rings are applied on tarsus, MRI: 2, while metal ring position is on tarsus or above or unknown, MRI: 1. However, the value 2 is not allowed by SOVON (see [comment in issue 47](https://github.com/inbo/cr-birding/issues/47#issuecomment-478926426)). Values allowed: 1, 4 and 5, where 4 is used for adding a color and/or metal ring, while 5 for changing

Mapping decision rules:

1. Default: MRI 4 
2. Very first ringing of a bird (no matter if `rngme` only, `rngkl` only or  `rngme` +  `rngkl`): `MRI` 1
3. Changing a metal and/or color ring: MRI 5

This last condition holds true even if the color ring is changed while adding a metal ring or viceversa.

We initialize `sovon_MRI` by assigning default value 4:

```{r initialize_MRI_obs_and_acts}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_MRI = 4)
```

Very first ringing gets `sovon_MRI` = 1:

```{r assign_1_to_very_first_ringing}
obs_and_acts <-
  obs_and_acts %>%
  group_by(sovon_bird_reference) %>%
  mutate(sovon_MRI = if_else(
    Datum == min(Datum) & (rngkl == "rngkl" | rngme == "rngme"),
    1, sovon_MRI
  )) %>%
  arrange(sovon_bird_reference, Datum) %>%
  ungroup()
```

Preview:

```{r preview_MRI_1}
obs_and_acts %>%
  filter(sovon_MRI == 1) %>%
  select(
    sovon_observation_reference,
    sovon_bird_reference,
    sovon_observation_date,
    sovon_MRI,
    acts
  ) %>%
  select_if(function(x) any(!is.na(x))) %>%
  head(n = 100)
```

Adding a color ring while metal ring is alrady present is mapped as 4, which is the default value. Same for adding a metal ring while color ring already present.

Based on `crbirding_birds` MRI is 5 if previous **`bird_shorthand` is not empty and different from actual `bird_shorthand` and no `klweg` action is coupled to the observation**.
As said in previous chapter while mapping ring data, we don't have track of changing metal rings, only loosing it (actions `meweg`).  We then assume to add a new one at next catching action. But adding a metal ring while not changing color ring means MRI equal to 4, default value, if color ring is changed then it is 5: this demonstrates that MRI value 5 depends only on value of `bird_shorthand` as only color rings can be changed in our mapping. If the ringer, while catching, realizes the absence of color ring (`klweg`), then the color ring is added, not changed, i.e. MRI equal 4 as metal ring is still present.If metal ring is absent as well, then it would be impossible to identify the bird and it would count as a new bird with an observation coupled to `rngkl` +  `rngme` actions.

```{r map_5_change_color_rings}
MRI_5_bird_ref_and_dates <-
  crbirding_birds %>%
  group_by(bird_reference) %>%
  mutate(previous_bird_shorthand = lag(bird_shorthand)) %>%
  filter(!is.na(previous_bird_shorthand) &
    previous_bird_shorthand != bird_shorthand) %>%
  select(bird_reference, bird_date_begin) %>%
  ungroup() %>%
  mutate(set_MRI_5 = 5)
obs_and_acts <-
  obs_and_acts %>%
  left_join(MRI_5_bird_ref_and_dates,
    by = c(
      "sovon_bird_reference" = "bird_reference",
      "Datum" = "bird_date_begin"
    )
  ) %>%
  mutate(sovon_MRI = if_else(!is.na(set_MRI_5) & is.na(klweg) & is.na(veld),
    set_MRI_5,
    sovon_MRI
  ))
```

Preview changes:

```{r preview_sovon_MRI_5}
obs_and_acts %>%
  filter(sovon_MRI == 5) %>%
  select(
    sovon_observation_reference,
    sovon_bird_reference,
    sovon_observation_date,
    sovon_MRI,
    acts
  ) %>%
  select_if(function(x) any(!is.na(x)))
```

The presence of action `ziek` in combination with `sovon_MRI` 5 is due to the three rings discussed in previous chapter: `FHOV`, `YCAF`, `FAAG`. 

### Bird age

The field `sovon_bird_age_obs` should be filled with the age of the bird as mentioned by the observer. This field has been created in previous chapter:

```{r mapping_bird_age_ringing}
obs_and_acts %>%
  distinct(sovon_bird_age_obs)
```

### Bird sex

The field `sovon_bird_sex` should be filled with the sex of the bird as mentioned by the observer. This field is not present in `obs_and_acts`, so we assign value `U` (unknown):

```{r add_sovon_bird_sex_obs}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_bird_sex = "U")
```

### Capture

The field `sovon_observation_is_capture` can be filled by evaluating the link of observations to actions `vang`, `vangl`, `rngkl` or `rngme`:

```{r is_capture}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_is_capture = if_else(
    !is.na(vang) | !is.na(vangl) | !is.na(rngkl) | !is.na(rngme), "Y", "N"
  ))
```

### Bird ID

The field `sovon_bird_id` is left to SOVON:

```{r add_bird_id}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_bird_id = NA)
```

### Bird ring position and inscription direction

Similarly to the mapping in ring data, we map the ring position and inscription reading direction in observations:

```{r map_pos_direction_obs}
obs_and_acts <-
  obs_and_acts %>%
  left_join(ring_position_table %>%
    select(Code, sovon_bird_ring_position),
  by = c("KleurringPlaats" = "Code")
  ) %>%
  left_join(ring_position_table %>%
    select(Code, sovon_bird_ring_direction),
  by = c("KleurringPlaats" = "Code")
  )
```

Effects of the mapping:

```{r result_mapping_position_direction_obs}
obs_and_acts %>%
  distinct(KleurringPlaats, sovon_bird_ring_position, sovon_bird_ring_direction)
```

### Observation notes

We copy the notes in `Opmerking` to SOVON field  `sovon_observation_notes`:

```{r obs_notes}
obs_and_acts <-
  obs_and_acts %>%
  mutate(sovon_observation_notes = Opmerking)
```

We add a dot at the end of the notes if not present already:

```{r add_final_dot}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_notes =
      if_else(
        !is.na(sovon_observation_notes) & str_length(sovon_observation_notes) > 0,
        if_else(str_sub(sovon_observation_notes, -1) != ".",
          str_c(sovon_observation_notes, ".", sep = ""),
          sovon_observation_notes
        ),
        sovon_observation_notes
      )
  )
```

We also add the prefix `"INBO original notes: "`:

```{r add_prefix}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_notes =
      if_else(
        !is.na(sovon_observation_notes) &
          str_length(sovon_observation_notes) > 0,
        str_c("INBO original notes: ", sovon_observation_notes, sep = " "),
        sovon_observation_notes
      )
  )
```

We add the note `no_color_ring.` to observations linked to action `klweg` and/or `me`:

```{r add_note_no_color_ring}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_notes = if_else(
      is.na(klweg) & is.na(me),
      sovon_observation_notes,
      if_else(is.na(sovon_observation_notes),
        "no_color_ring.",
        str_c("no_color_ring.", sovon_observation_notes, sep = " ")
      )
    )
  )
```

Preview:

```{r no_color_ring_example}
obs_and_acts %>%
  filter(!is.na(klweg) | !is.na(me)) %>%
  select(, klweg, me, sovon_observation_notes) %>%
  head()
```

We add the note `no_metal_ring.` to observations linked to action `meweg`:

```{r add_note_no_metal_ring}
obs_and_acts <-
  obs_and_acts %>%
  mutate(
    sovon_observation_notes = if_else(
      !is.na(meweg),
      if_else(is.na(sovon_observation_notes),
        "no_metal_ring.",
        paste("no_metal_ring.", sovon_observation_notes, sep = " ")
      ),
      sovon_observation_notes
    )
  )
```

Preview:

```{r no_metal_ring_example}
obs_and_acts %>%
  filter(!is.na(meweg)) %>%
  select(sovon_observation_notes) %>%
  distinct() %>%
  head()
```

## Save final observation data

Select the required columns, starting with prefix `sovon_`:

```{r select_sovon_cols}
crbirding_observations <-
  obs_and_acts %>%
  select(starts_with("sovon"))
```

Remove prefix `sovon_`:

```{r remove prefix_sovon_observations}
names(crbirding_observations) <- str_remove_all(names(crbirding_observations), pattern = "sovon_")
```

Remove time and timezone information from `bird_date_begin` and `bird_date_end`:

```{r remove_time_and_tz_obs}
crbirding_observations <-
  crbirding_observations %>%
  mutate(observation_date = as.Date(observation_date))
```

The desired order of columns in `crbirding_observations`:

```{r cols_order_obs}
cr_obs_cols <- c(
  "user_id", "user_reference", "bird_id", "bird_reference", "observation_id",
  "observation_reference", "observation_date", "observation_time",
  "observation_lat", "observation_lng", "observation_location",
  "observation_is_capture", "observation_notes", "check_bird", "MRI", "melder",
  "melder_email", "melder_ringersnummer", "ring_number", "observation_condition", "status_full_grown_bird",
  "bird_age_obs", "bird_sex", "bird_ring_position", "bird_ring_direction"
)
```

Are all required columns present?

```{r check_presence_required_cols_obs}
all(cr_obs_cols %in% names(crbirding_observations)) &
  length(cr_obs_cols) == ncol(crbirding_observations)
```

Set column order:

```{r get_right_order_cols_obs}
crbirding_observations <-
  crbirding_observations %>%
  select(all_of(cr_obs_cols))
```

Preview data (without personal observer data for privcay reasons):

```{r final_preview_obs}
crbirding_observations %>%
  select(-c(melder, melder_email)) %>%
  head(n = 10)
```

Save to text file (comma separated value):

```{r write_processed_obs_data}
write_csv(
  crbirding_observations,
  path = here::here("data", "processed", "crbirding_observations.csv"),
  na = ""
)
```
