==> src/5_map_to_previous_output.Rmd <==
# Compare to data transfer of 2019 and preserve IDs

SOVON asks us to preserve IDs used in mapped data as sent in April 2019.

## Read data

First, we need to read the data we sent them in April 2019.

```{r read_2019_data}
crbirding_users_2019 <- read_csv(here::here("data",
                                  "processed",
                                  "crbirding_users_20190410.csv"),
                                 col_types = cols(
                                   .default = col_character(),
                                   user_id = col_logical(),
                                   user_reference = col_number()),
                       na = "")
crbirding_birds_2019 <- read_csv(here::here("data",
                                  "processed",
                                  "crbirding_birds_20190410.csv"),
                       col_types = cols(
                         .default = col_character(),
                         bird_id = col_logical(),
                         bird_reference = col_number(),
                         bird_bto = col_logical(),
                         bird_birth_year = col_logical(),
                         bird_date_begin = col_datetime(
                           format = "%d-%m-%Y"),
                         bird_date_end = col_datetime(
                           format = "%d-%m-%Y")),
                       na = "")
crbirding_observations_2019 <- read_csv(here::here(
  "data",
  "processed",
  "crbirding_observations_20190410.csv"),
  col_types = cols(
    .default = col_character(),
    user_id = col_logical(),
    user_reference = col_number(),
    bird_id = col_logical(),
    bird_reference = col_number(),
    observation_id = col_logical(),
    observation_reference = col_number(),
    observation_date = col_datetime(
      format = "%d-%m-%Y"),
    observation_time = col_logical(),
    check_bird = col_logical(),
    MRI = col_number(),
    melder_ringersnummer = col_logical(),
    condition = col_number()
  ),
  na = "")
```

We read also the data we mapped just now.

```{r read_data_now}
crbirding_users <- read_csv(here::here("data",
                                       "processed",
                                       "crbirding_users.csv"),
                            col_types = cols(
                              .default = col_character(),
                              user_id = col_logical(),
                              user_reference = col_number()),
                            na = "")
crbirding_birds <- read_csv(here::here("data",
                                       "processed",
                                       "crbirding_birds.csv"),
                            col_types = cols(
                              .default = col_character(),
                              bird_id = col_logical(),
                              bird_reference = col_number(),
                              bird_bto = col_logical(),
                              bird_birth_year = col_logical(),
                              bird_date_begin = col_datetime(
                                format = ""),
                              bird_date_end = col_datetime(
                                format = "")),
                            na = "")
crbirding_observations <- read_csv(here::here(
  "data",
  "processed",
  "crbirding_observations.csv"),
  col_types = cols(
    .default = col_character(),
    user_id = col_logical(),
    user_reference = col_number(),
    bird_id = col_logical(),
    bird_reference = col_number(),
    observation_id = col_logical(),
    observation_reference = col_number(),
    observation_date = col_datetime(
      format = ""),
    observation_time = col_logical(),
    check_bird = col_logical(),
    MRI = col_number(),
    melder_ringersnummer = col_logical(),
    observation_condition = col_number()
  ),
  na = "")
```

To improve the comparison of bird data, we read also raw observations:

```{r import_obs_and_acts}
obs_and_acts <- read_tsv(
  here::here("data", "interim", "obs_and_actions.tsv"),
  col_types = cols(
    .default = col_character(),
    sovon_bird_reference = col_double(),
    Nummer = col_double(),
    Datum = col_datetime(format = ""),
    WaarnemerNummer = col_double(),
    PlaatsLengtegraadGraden = col_double(),
    PlaatsLengtegraadMinuten = col_double(),
    PlaatsLengtegraadSeconden = col_double(),
    PlaatsBreedtegraadGraden = col_double(),
    PlaatsBreedtegraadMinuten = col_double(),
    PlaatsBreedtegraadSeconden = col_double(),
    BevestigingDatum = col_datetime(format = ""),
    AanmaakDatum = col_datetime(format = ""),
    WijzigDatum = col_datetime(format = "")
  )
)
```


## Users

Number of new users based on new `user_reference` (email and other private info not shown):

```{r new_user_reference}
crbirding_users %>%
  filter(!user_reference %in% crbirding_users_2019$user_reference) %>%
  select(user_reference, user_country, user_role)
```

Old `user_reference` IDs not anymore present (email and other private info not shown):

```{r old_user_ref_not_present}
crbirding_users_2019 %>%
  filter(!user_reference %in% crbirding_users$user_reference) %>%
  select(user_reference, user_country, user_role)
```

Users from 2019 where some changes have been applied:

```{r users_with_changes}
same_ref_ids <- 
  crbirding_users %>%
  filter(user_reference %in% crbirding_users_2019$user_reference) %>%
  pull(user_reference)
ids_users_with_changes <- 
  crbirding_users_2019 %>%
  filter(user_reference %in% same_ref_ids) %>%
  anti_join(crbirding_users,
            by = names(crbirding_users)) %>%
  pull(user_reference)
ids_users_with_changes
```

## Map birds

Before mapping birds, we have to modify the color rings in ` crbirding_birds_2019` by applying the syntax discussed in [#107](https://github.com/inbo/cr-birding/issues/107) and [#67](https://github.com/inbo/cr-birding/issues/67). SOVON asks to follow this syntax: ring color + hyphen (`-`) + inscription without inscription color, e.g. `W-EAAM` instead of `WN(EAAM)`.

First, check that all color rings from 2019 have two letters before bracket `(`:

```{r check_format_2019}
crbirding_birds_2019 %>%
  filter(is.na(bird_shorthand) | 
           str_sub(bird_shorthand, start = 3, end = 3) == "(") %>%
  nrow == nrow(crbirding_birds_2019)
```

We proceed then to modify the color rings:

```{r modify_color_rings_2019}
crbirding_birds_2019 <-
  crbirding_birds_2019 %>%
  mutate(bird_shorthand = if_else(
    !is.na(bird_shorthand),
    str_c(str_sub(bird_shorthand, start = 1, end = 1),
          "-",
          str_extract(bird_shorthand, pattern = "(?<=\\()(.*?)(?=\\))")),
    bird_shorthand
))
```

Preview:

```{r preview_color_rings_2019_new_syntax}
crbirding_birds_2019 %>%
  select(bird_shorthand) %>%
  filter(!is.na(bird_shorthand)) %>%
  head(n = 100)
```

Now we can proceed mapping birds data.

### Map `bird_reference`

In INBO ring table, birds were uniquely identified by their very first ring (data were spread and untidy). We added a `bird_reference` field to gather information about multiple rings and link them to the same bird. This means that this field cannot be used to compare the actual INBO data with the data of 2019. We have to compare data by the very first color ring.

As the dot for readibility is only preserved for the very last ring, it can be better to compare data by removing it from bird shorthand. We create an auxiliary column called `inscription` for both `crbirding_birds` and  `crbirding_birds_2019`:

```{r add_inscription}
crbirding_birds_2019 <- 
  crbirding_birds_2019 %>%
  mutate(inscription = if_else(!is.na(bird_shorthand),
                               str_extract(
                                str_remove(bird_shorthand, "\\."),
                                "(?<=\\-)[A-Z0-9]*"),
                              NA_character_))
crbirding_birds <- 
  crbirding_birds %>%
  mutate(inscription = if_else(!is.na(bird_shorthand),
                               str_extract(
                                str_remove(bird_shorthand, "\\."),
                                "(?<=\\-)[A-Z0-9]*"),
                              NA_character_))
```

We add the first color ring inscription as additional column, `first_color_ring`:

```{r first_color_ring}
crbirding_birds_first_color_ring <-
  crbirding_birds %>% 
  select(bird_reference, inscription, bird_date_begin) %>%
  group_by(bird_reference) %>%
  filter(!is.na(inscription)) %>%
  mutate(first_color_ring = if_else(
    bird_date_begin == min(bird_date_begin, na.rm = TRUE),
    inscription,
    NA_character_)) %>%
  ungroup() %>%
  filter(!is.na(first_color_ring)) %>%
  distinct(bird_reference, first_color_ring) %>%
  right_join(crbirding_birds, by = c("bird_reference"))
```

All birds should get a valid `first_color_ring` column. Exceptions:

```{r birds_without_first_color_ring}
crbirding_birds_first_color_ring %>%
  filter(is.na(first_color_ring))
```

We do the same with ring data from 2019:

```{r first_color_ring_2019}
crbirding_birds_first_color_ring_2019 <-
  crbirding_birds_2019 %>% 
  select(bird_reference, inscription, bird_date_begin) %>%
  group_by(bird_reference) %>%
  filter(!is.na(inscription)) %>%
  mutate(first_color_ring = if_else(
    bird_date_begin == min(bird_date_begin,na.rm = TRUE),
    inscription,
    NA_character_)) %>%
  ungroup() %>%
  filter(!is.na(first_color_ring)) %>%
  distinct(bird_reference, first_color_ring) %>%
  right_join(crbirding_birds_2019, by = c("bird_reference"))
```

Even in this case all birds should get a valid `first_color_ring` column. Exceptions:

```{r birds_without_first_color_ring_2019}
crbirding_birds_first_color_ring_2019 %>%
  filter(is.na(first_color_ring))
```

Now we can map the new `bird_reference` to the `bird_reference` of 2019:

```{r mapping_bird_reference_new_old}
crbirding_birds_mapping_bird_reference <-
  crbirding_birds_first_color_ring %>%
  distinct(bird_reference, first_color_ring) %>%
  left_join(crbirding_birds_first_color_ring_2019 %>%
              distinct(bird_reference, first_color_ring) %>%
              rename(bird_reference_2019 = bird_reference),
            by = "first_color_ring") %>%
  select(first_color_ring, everything())
head(crbirding_birds_mapping_bird_reference, n = 100)
```

We add `bird_reference_2019` to `crbirding_birds` and `crbirding_observations`:

```{r add_bird_ref_2019}
crbirding_birds <- 
  crbirding_birds %>%
  left_join((crbirding_birds_mapping_bird_reference %>%
              select(bird_reference, bird_reference_2019)),
            by = "bird_reference") %>%
  select(bird_id, bird_reference, bird_reference_2019, everything())
```

and `crbirding_observations`:

```{r add_bird_ref_2019_to_obs}
crbirding_observations <-
  crbirding_observations %>%
  left_join((crbirding_birds_mapping_bird_reference %>%
              select(bird_reference, bird_reference_2019)),
            by = "bird_reference")
```
 
Number of birds present in 2019 but with `bird_reference` not equal to `bird_reference_2019` :

```{r n_birds_changed_bird_reference}
crbirding_birds %>%
  filter(!is.na(bird_reference_2019)) %>%
  filter(bird_reference != bird_reference_2019) %>%
  nrow
```

This is not surprising as `bird_reference` has been defined as a progressive number, so it is sufficient the change of one bird reference to modify all the other references coming after.

First change appears at:

```{r first_change}
crbirding_birds %>%
  filter(!is.na(bird_reference_2019)) %>%
  filter(bird_reference != bird_reference_2019) %>%
  filter(bird_reference == min(bird_reference)) %>%
  select(bird_reference, bird_reference_2019, bird_shorthand)
```

### New `bird_reference`

Some birds were not present in data transfer of April 2019, so the `bird_reference` cannot be mapped to `bird_reference_2019`.

This case includes:

1. New birds: birds ringed for the very first time after the previous data transfer in April 2019.
2. Birds with modified first color ring: birds whose very first color ring has been corrected or added. This case is due to the fact that the first color ring is the real unique identifier in INBO ring table `tblKleurring`

```{r new_birds_or_modified_first_ring}
crbirding_birds %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference_2019" = "bird_reference"))
```

Based on column `bird_date_begin` (see above) we can see that most but not all rings have been placed after 2019-04-10, date of previous export. It can also occur that there are old rings which have been added to database after that date, so they are actually _new_ rings, database speaking. So, to check which rings have been entered before 2019-04-10 we use column `AanmaakDatum` in raw observation data:

```{r supected_new_birds_before_export_date}
supected_new_birds_before_export_date <- 
  crbirding_birds %>%
  filter(is.na(bird_reference_2019)) %>%
  left_join(crbirding_birds_first_color_ring %>%
              distinct(bird_reference, first_color_ring),
            by = "bird_reference") %>%
  select(bird_reference, first_color_ring) %>%
  left_join(obs_and_acts,
            by = c("first_color_ring" = "KleurringNummer")) %>%
  select(bird_reference, first_color_ring, AanmaakDatum) %>%
  group_by(bird_reference, first_color_ring) %>%
  filter(AanmaakDatum == min(AanmaakDatum)) %>%
  ungroup() %>%
  filter(AanmaakDatum < as_datetime("2019-04-10")) %>%
  left_join(crbirding_birds,
            by = c("bird_reference"))
supected_new_birds_before_export_date
```

To better retrieve potential matches to birds from 2019's export, we check birds from 2019 not included in actual data:

```{r birds_2019_not_in_new_export}
crbirding_birds_first_color_ring_2019 %>%
  anti_join(crbirding_birds_first_color_ring,
            by = "first_color_ring")
```

As expected most of these rings are the so called virtual rings that we have decided to not include in new export (see [#81-issuecomment-698964036]( https://github.com/inbo/cr-birding/issues/81#issuecomment-698964036)).

If we exclude them from this overview, the following birds remain:

```{r birds_2019_not_in_new_export_except_virtual_rings}
crbirding_birds_first_color_ring_2019 %>%
  filter(str_sub(first_color_ring, start = 3, end = 3) != "B") %>%
  anti_join(crbirding_birds_first_color_ring,
            by = "first_color_ring")
```

The ring `ALOB` is not the first ring in the new data, but it replaces the ring `LPAV`. We match it and assign `bird_reference_2019` to this ring:

```{r assign_bird_ref_2019_to_ALOB}
bird_reference_ALOB <- 
  crbirding_birds %>%
  filter(inscription == "ALOB") %>%
  pull(bird_reference)
bird_reference_ALOB_2019 <-
  crbirding_birds_2019 %>%
  filter(inscription == "ALOB") %>%
  pull(bird_reference)
crbirding_birds <-
  crbirding_birds %>%
  mutate(bird_reference_2019 = if_else(bird_reference == bird_reference_ALOB,
                                       bird_reference_ALOB_2019,
                                       bird_reference_2019))
crbirding_observations <-
  crbirding_observations %>%
  mutate(bird_reference_2019 = if_else(bird_reference == bird_reference_ALOB,
                                       bird_reference_ALOB_2019,
                                       bird_reference_2019))
```

The ring `EAU` has been removed from the database at the preprocessing stage (see section about the [removal of some rings](#remove-three-rings)).

Based on metal ring (`bird_ring_number`) information, date of ringing (`bird_date_begin`) and other fields we can see that the ring FMOV is the same as ring FMOU from 2019 export:

```{r check_FMOV-FMOU}
all(
  crbirding_birds_2019 %>%
    filter(bird_shorthand == "B-FM.OU") %>%
    select(bird_date_begin, bird_ring_number, bird_age_ringing, bird_ring_position) %>%
    pull() ==
    (crbirding_birds %>%
       filter(bird_shorthand == "B-FM.OV") %>%
       select(bird_date_begin, bird_ring_number, bird_age_ringing, bird_ring_position) %>%
       pull()))
```

So, we can assign `bird_reference_2019` to this ring:

```{r assign_bird_ref_2019_to_FMOV}
bird_reference_FMOV <- 
  crbirding_birds %>%
  filter(inscription == "FMOV") %>%
  pull(bird_reference)
bird_reference_FMOU_2019 <- 
  crbirding_birds_2019 %>%
  filter(inscription == "FMOU") %>%
  pull(bird_reference)
crbirding_birds <-
  crbirding_birds %>%
  mutate(bird_reference_2019 = if_else(bird_reference == bird_reference_FMOV,
                                       bird_reference_FMOU_2019,
                                       bird_reference_2019))
crbirding_observations <-
  crbirding_observations %>%
  mutate(bird_reference_2019 = if_else(bird_reference == bird_reference_FMOV,
                                       bird_reference_FMOU_2019,
                                       bird_reference_2019))
```

Show changes in column `bird_reference_2019`:

```{r changes_FMOV_ALOB}
crbirding_birds %>%
  filter(bird_reference %in% c(bird_reference_ALOB, bird_reference_FMOV)) %>%
  select(bird_reference, bird_reference_2019, bird_shorthand)
```

### Use `bird_reference` from 2019 as primary key

To ease the work of SOVON It team and as discussed in [issue 128](https://github.com/inbo/cr-birding/issues/128) we reuse the old bird references (`bird_reference_2019`) as primary key in `crbirding_birds` where present, being careful to change the foreign key in `crbirding_observations` as well.

We do it by using a helping column, `new_bird_reference` filled with old bird references where possible and with the new ones plus 20000 to ensure unicity:

```{r add_new_bird_reference}
crbirding_birds <-
  crbirding_birds %>%
  mutate(new_bird_reference = if_else(!is.na(bird_reference_2019),
                                      bird_reference_2019,
                                      bird_reference + 20000))
crbirding_observations <-
  crbirding_observations %>%
  mutate(new_bird_reference = if_else(!is.na(bird_reference_2019),
                                      bird_reference_2019,
                                      bird_reference + 20000))

```

Is `new_bird_reference` unique for both birds and observations?

```{r unicity_new_bird_reference}
nrow(crbirding_birds %>%
  distinct(bird_reference, new_bird_reference)) == 
  nrow(crbirding_birds %>%
         distinct(bird_reference))
nrow(crbirding_observations %>%
  distinct(bird_reference, new_bird_reference)) == 
  nrow(crbirding_observations %>%
         distinct(bird_reference))
```

Again, double checking that `new_bird_reference` is unique:

```{r check_unicity_fk_obs}
nrow(crbirding_observations %>%
  distinct(bird_reference, new_bird_reference)) == 
  nrow(crbirding_observations %>%
         distinct(bird_reference))
```

Final check: are all bird reference ids in `new_bird_reference` in `crbirding_observations` in `crbirding_birds` as well?

```{r check_all_bird_ref_id_in_obs}
crbirding_observations %>%
  filter(!new_bird_reference %in% crbirding_birds$new_bird_reference) %>%
  nrow() == 0
```

If `FALSE`, it means some observations have a broken link to birds and action is needed to solve this bug.

If `TRUE`, we can safely copy the values of `new_bird_reference` to `bird_reference` and remove the help columns `bird_reference_2019` and `inscription`:

```{r set_bird_ref_again}
crbirding_birds <- 
  crbirding_birds %>%
  select(-c(bird_reference, bird_reference_2019, inscription)) %>%
  rename(bird_reference = new_bird_reference) %>%
  # reorder columns 
  select(bird_id, bird_reference, everything())

crbirding_observations <- 
  crbirding_observations %>%
  select(-bird_reference) %>%
  rename(bird_reference = new_bird_reference) %>%
  # reorder columns 
  select(user_id,
         user_reference,
         bird_id,
         bird_reference,
         everything())

crbirding_birds_2019 <- 
  crbirding_birds_2019 %>%
  select(-c(inscription))
```

### Changed  `bird_euring` 

```{r changed_bird_euring}
crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference",
                   "bird_euring")) %>%
  select(bird_reference, bird_euring) %>%
  left_join(crbirding_birds_2019 %>%
              rename(bird_euring_2019 = bird_euring) %>%
              select(bird_reference, bird_euring_2019),
            by = ("bird_reference"))
```

### Changed  `bird_shorthand` 

Birds with changed ring history:

```{r changed_bird_shorthand}
bird_ref_changed_shorthand <-
  crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  select(bird_reference, bird_shorthand) %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference", "bird_shorthand")) %>%
  distinct(bird_reference)
bird_ref_changed_shorthand
```

`bird_shorthand` history in  `crbirding_birds`:

```{r different_bird_shorthand}
crbirding_birds %>%
  filter(bird_reference %in% bird_ref_changed_shorthand$bird_reference) %>%
  select(bird_reference,
         bird_shorthand,
         bird_date_begin,
         bird_date_end) %>%
  arrange(bird_reference)
```

Values of `bird_shorthand` in  `crbirding_birds_2019`:

```{r bird_shorthand_in_2019}
crbirding_birds_2019 %>%
  filter(bird_reference %in% bird_ref_changed_shorthand$bird_reference) %>%
  select(bird_reference,
         bird_shorthand,
         bird_date_begin,
         bird_date_end) %>%
  # add "_2019" to make sure we are speaking of 2019's export data
  rename_all(paste0, "_2019") %>%
  arrange(bird_reference_2019)
```

Most of these changes are due to change the color rings.

### Changed `bird_scheme`

Rings with changed `bird_scheme`:

```{r changed_bird_scheme}
changed_bird_scheme <- 
  crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference", "bird_scheme")) %>%
  select(bird_reference, bird_reference,
         bird_scheme, bird_date_begin, bird_date_end) %>%
  left_join(crbirding_birds_2019 %>%
              rename(bird_scheme_2019 = bird_scheme) %>%
              select(bird_reference,
                     bird_scheme_2019,
                     bird_date_begin,
                     bird_date_end),
            by = (c("bird_reference", "bird_date_begin", "bird_date_end"))) %>%
  filter(!is.na(bird_scheme_2019) | !is.na(bird_scheme)) %>%
  select(bird_reference,
         bird_scheme,
         bird_scheme_2019,
         everything())
changed_bird_scheme
```

Mapping values:

```{r mapping_changes_bird_scheme}
changed_bird_scheme %>%
  group_by(bird_scheme, bird_scheme_2019) %>%
  count()
```

### Changed `bird_ring_number`

Rings with changed `bird_ring_number`:

```{r changed_bird_ring_number}
changed_bird_ring_number <- 
  crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference", "bird_ring_number")) %>%
  select(bird_reference, bird_ring_number, bird_date_begin, bird_date_end) %>%
  left_join(crbirding_birds_2019 %>%
              rename(bird_ring_number_2019 = bird_ring_number) %>%
              select(bird_reference,
                     bird_ring_number_2019,
                     bird_date_begin,
                     bird_date_end),
            by = (c("bird_reference", "bird_date_begin", "bird_date_end"))) %>%
  filter(!is.na(bird_ring_number_2019) | !is.na(bird_ring_number)) %>%
  select(contains("reference"),
         contains("number"),
         contains("date"))
changed_bird_ring_number
```

### Changed `bird_name`

Rings with changed `bird_name`:

```{r changed_bird_name}
changed_bird_name <- 
  crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference", "bird_name")) %>%
  select(bird_reference, 
         bird_name, 
         bird_date_begin,
         bird_date_end) %>%
  left_join(crbirding_birds_2019 %>%
              rename(bird_name_2019 = bird_name) %>%
              select(bird_reference,
                     bird_name_2019,
                     bird_date_begin,
                     bird_date_end),
            by = (c("bird_reference", "bird_date_begin", "bird_date_end"))) %>%
  filter(!is.na(bird_name_2019) | !is.na(bird_name)) %>%
  select(bird_reference,
         contains("name"),
         contains("date"))
changed_bird_name
```

### Changed `bird_sex`

Rings with changed `bird_sex`:

```{r changed_bird_sex}
changed_bird_sex <- 
  crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  anti_join(crbirding_birds_2019,
            by = c("bird_reference", "bird_sex")) %>%
  select(bird_reference, 
         bird_sex, 
         bird_date_begin,
         bird_date_end) %>%
  left_join(crbirding_birds_2019 %>%
              rename(bird_sex_2019 = bird_sex) %>%
              select(bird_reference,
                     bird_sex_2019,
                     bird_date_begin,
                     bird_date_end),
            by = (c("bird_reference", "bird_date_begin", "bird_date_end"))) %>%
  filter(!is.na(bird_sex_2019) | !is.na(bird_sex)) %>%
  select(contains("reference"),
         contains("sex"),
         contains("date"))
changed_bird_sex
```

### Changed `bird_date_begin`

Rings with changed `bird_date_begin` excluding the birds with new rings:

```{r changed_bird_date_begin}
same_rings <- 
  crbirding_birds %>%
  # remove new rings
  filter(bird_reference - 20000 < 0) %>%
  select(bird_reference,
         bird_shorthand,
         bird_date_begin,
         bird_date_end) %>%
  inner_join(crbirding_birds_2019 %>%
               select(bird_reference,
                      bird_date_begin_2019 = bird_date_begin,
                      bird_shorthand,
                      bird_date_end_2019 = bird_date_end),
             by = c("bird_reference", "bird_shorthand"))
changed_bird_date_begin_end <-
  same_rings %>%
  group_by(bird_reference) %>%
  filter(!bird_date_begin %in% bird_date_begin_2019 | 
           !bird_date_end %in% bird_date_end_2019) %>%
  filter(!is.na(bird_date_begin) &
           !is.na(bird_date_begin_2019) &
           !is.na(bird_date_end) &
           !is.na(bird_date_end_2019)) %>%
  distinct(bird_reference)
changed_bird_date_begin_end
```

Actual color ring history:

```{r actual_history_changed_rings}
changed_bird_date_begin_end %>%
  left_join(crbirding_birds,
            by = "bird_reference") %>%
  select(bird_reference, bird_shorthand, bird_date_begin, bird_date_end)
```

Color ring history data from 2019:

```{r history_changed_rings_from_2019}
changed_bird_date_begin_end %>%
  left_join(crbirding_birds_2019,
            by = "bird_reference") %>%
  select(bird_reference, bird_shorthand, bird_date_begin, bird_date_end)
```

### Save bird data with updated `bird_reference`

We save `crbirding_birds` and `crbirding_observations` with the updated unique identifiers in `bird_reference`:

```{r save_birds_with_bird_ref_2019}
# save bird data
crbirding_birds <- 
  crbirding_birds %>%
  mutate(
    bird_date_begin = as.Date(bird_date_begin),
    bird_date_end = as.Date(bird_date_end)) %>%
  write_csv(path = here::here("data",
                              "processed",
                              "crbirding_birds.csv"),
  na = ""
)
```

## Map observations

Number of new observations based on new `observation_reference` (email and other private info not shown):

```{r new_observation_reference}
crbirding_observations %>%
  filter(!observation_reference %in% 
           crbirding_observations_2019$observation_reference) %>%
  nrow()
```

Old observations not anymore present (all columns info from 2019 data):

```{r old_obs_ref_not_present}
crbirding_observations_2019 %>%
  filter(!observation_reference %in% 
           crbirding_observations$observation_reference) %>%
  select(-contains("melder"))
```

Save in `same_ref_ids` the observation reference of observations present in both data.frames:

```{r same_ref_ids_obs}
same_ref_ids <- 
  crbirding_observations %>%
  filter(observation_reference %in% crbirding_observations_2019$observation_reference) %>%
  pull(observation_reference)
```

### Changed geospatial information

Details of observations with changes in latitude (`observation_lat`) and/or longitude (`observation_lng`):

```{r obs_with_changes_lat_lng}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "observation_lat",
                   "observation_lng")) %>%
  select(observation_reference, observation_lat, observation_lng) %>%
  rename(observation_lat_2019 = observation_lat,
         observation_lng_2019 = observation_lng) %>%
  left_join(crbirding_observations %>%
              select(observation_reference,
                     observation_lat,
                     observation_lng),
            by = "observation_reference") %>%
  select(observation_reference,
         observation_lat_2019,
         observation_lat,
         observation_lng_2019,
         observation_lng)
```

Notice how these are observations where the longitude direction was not mapped properly during export of 2019 (see issue [#105](https://github.com/inbo/cr-birding/issues/105)).

Details of observations with changes in location (column `observation_location`):

```{r obs_with_changes_location}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "observation_location")) %>%
  select(observation_reference, observation_location) %>%
  rename(observation_location_2019 = observation_location) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, observation_location),
            by = "observation_reference")
```

### Changed temporal information

Details of observations with changes in temporal information (column `observation_date`):

```{r obs_with_changes_date}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "observation_date")) %>%
  select(observation_reference, observation_date) %>%
  rename(observation_date_2019 = observation_date) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, observation_date),
            by = "observation_reference")
```

### Changed user information

Details of observations with changes in user detail(no details about users shown, just `user_reference`):

```{r obs_with_changes_user}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "user_reference",
                   "melder",
                   "melder_email")) %>%
  select(observation_reference, user_reference) %>%
  rename(user_reference_2019 = user_reference) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, user_reference),
            by = "observation_reference")
```

Most `user_reference` are the same as in 2019: it means that changes are in the columns hidden for privacy reasons. These user ids are indeed the ones stored above in variable `ids_users_with_changes` in the [Users] section.

### Changed metal ring information

Details of observations with changes in metal ring information (columns `MRI` and  `ring_number`):

```{r obs_with_changes_metalring_infos}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "MRI",
                   "ring_number")) %>%
  select(observation_reference, MRI, ring_number) %>%
  rename(MRI_2019 = MRI,
         ring_number_2019 = ring_number) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, MRI, ring_number),
            by = "observation_reference")
```

### Changed condition

First we rename column `condition` of  `crbirding_observation_2019` to new column name `observation_condition`:

```{r rename_condition}
crbirding_observations_2019 <-
  crbirding_observations_2019 %>%
  rename(observation_condition = condition)
```

Details of observations with changes in condition:

```{r obs_with_changes_condition}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "observation_condition")) %>%
  select(observation_reference, observation_condition) %>%
  rename(observation_condition_2019 = observation_condition) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, observation_condition),
            by = "observation_reference")
```

A lot of changes expected as the mapping codes used for observation_condition changed.

### Changed status

First we rename column `status` of  `crbirding_observation_2019` to new column name `status_full_grown_bird`:

```{r rename_status}
crbirding_observations_2019 <-
  crbirding_observations_2019 %>%
  rename(status_full_grown_bird = status)
```

Details of observations with changes in status (column `status_full_grown_bird`):

```{r obs_with_changes_status}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "status_full_grown_bird")) %>%
  select(observation_reference, status_full_grown_bird) %>%
  rename(status_full_grown_bird_2019 = status_full_grown_bird) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, status_full_grown_bird),
            by = "observation_reference")
```

### Changed bird age

Details of observations with changes in bird age at time of observation (column `bird_age_obs`):

```{r obs_with_changes_bird_age_obs}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "bird_age_obs")) %>%
  select(observation_reference, bird_age_obs) %>%
  rename(bird_age_obs_2019 = bird_age_obs) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, bird_age_obs),
            by = "observation_reference")

```

### Changed bird sex

Details of observations with changes in bird sex at time of observation (column `bird_sex`):

```{r obs_with_changes_bird_sex}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "bird_sex")) %>%
  select(observation_reference, bird_sex) %>%
  rename(bird_sex_2019 = bird_sex) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, bird_sex),
            by = "observation_reference")

```

### Changed bird ring position

Details of observations with changes in bird ring position (column `bird_ring_position`) or direction (`bird_ring_direction`):

```{r obs_with_changes_bird_ring_position_direction}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference", 
                   "bird_ring_position",
                   "bird_ring_direction")) %>%
  select(observation_reference,
         bird_ring_position, 
         bird_ring_direction) %>%
  rename(bird_ring_position_2019 = bird_ring_position,
         bird_ring_direction_2019 = bird_ring_direction) %>%
  left_join(crbirding_observations %>%
              select(observation_reference,
                     bird_ring_position,
                     bird_ring_direction),
            by = "observation_reference") %>%
  select(observation_reference,
         bird_ring_position_2019,
         bird_ring_position,
         bird_ring_direction_2019,
         bird_ring_direction)
```

### Changed notes

Details of observations with changes in observation notes (column `observation_notes`):

```{r obs_with_changes_observation_notes}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "observation_notes")) %>%
  select(observation_reference, observation_notes) %>%
  rename(observation_notes_2019 = observation_notes) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, observation_notes),
            by = "observation_reference")
```

### Changed capture information

Details of observations with changes in capture information (column `observation_is_capture`):

```{r obs_with_changes_observation_is_capture}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("observation_reference",
                   "observation_is_capture")) %>%
  select(observation_reference, observation_is_capture) %>%
  rename(observation_is_capture_2019 = observation_is_capture) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, observation_is_capture),
            by = "observation_reference")
```

### Changed bird reference

Details of observations with changes in bird reference (column `bird_reference`):

```{r obs_with_changes_bird_reference}
crbirding_observations_2019 %>%
  filter(observation_reference %in% same_ref_ids) %>%
  anti_join(crbirding_observations,
            by = c("bird_reference",
                   "observation_reference")) %>%
  select(observation_reference, bird_reference) %>%
  rename(bird_reference_2019 = bird_reference) %>%
  left_join(crbirding_observations %>%
              select(observation_reference, bird_reference),
            by = "observation_reference")
```

The huge number of changes of `bird_reference` has been discussed in previous session related to `crbirding_birds`.

### Save observation data with updated `bird_reference`

We save observation data, `crbirding_observations`, with the updated bird unique identifiers, `bird_reference`:

```{r save_obs_and_obs_with_bird_ref_2019}
crbirding_observations <- 
  crbirding_observations %>%
  mutate(
    observation_date = as.Date(observation_date)) %>%
  write_csv(path = here::here("data",
                              "processed",
                              "crbirding_observations.csv"),
  na = ""
)
```
